% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{addAlleleCounts}
\alias{addAlleleCounts}
\title{Add allele counts to a \code{muscadet} object}
\usage{
addAlleleCounts(x, allele_counts)
}
\arguments{
\item{x}{A \code{\link{muscadet}} object.}

\item{allele_counts}{A list of data frames where each data frame contains
allele counts for a specific omic (\code{list}). The list must have the same
length and order as the number of omics in the \code{x} object. Each data frames
must contain the following columns : \code{cell}, \code{id}, \code{CHROM}, \code{POS}, \code{REF},
\code{ALT}, \code{RD}, \code{AD}, \code{DP}, \code{GT}. See \link{allele_counts} for details.}
}
\value{
A modified \code{\link{muscadet}} object with updated allele counts in the
\code{allelic} slot of each \code{muscomic} object in the \code{omics} slot.
}
\description{
This function adds allele counts data to the \code{omics} of a
\code{\link{muscadet}} object. The data frames in the \code{allele_counts} list
are assigned to the \code{allelic} slots of \code{omics}.
}
\note{
As the allele counts are not used for computing log R
ratios and cell clustering, they are not mandatory at the creation of
\code{muscomic} and \code{muscadet} objects. The allele counts data can be added to
objects later with this \code{addAlleleCounts} function, before using the
\code{\link[=mergeCounts]{mergeCounts()}} function.

This function is also useful to add allele counts for sample-specific SNP
positions to a common \code{muscadet} object, for example for the reference
\code{muscadet} object: a common \code{muscadet} object with reference coverage data
can be stored as a unique object to use as reference for computing log R
ratios for all samples, and then it can updated with allele counts at sample
specific SNP positions (e.g. found by bulk sequencing) before Copy Number
Alterations (CNAs) calling.
}
\examples{
# Load example muscadet objects
data(muscadet_obj)
data(muscadet_obj_ref)

# Add allele counts data frames to muscadet objects
muscadet_obj <- addAlleleCounts(
    muscadet_obj,
    allele_counts = list(allele_counts_atac_tumor, allele_counts_rna_tumor))
muscadet_obj_ref <- addAlleleCounts(
    muscadet_obj_ref,
    allele_counts = list(allele_counts_atac_ref, allele_counts_rna_ref))

}
\seealso{
\code{\link[=CreateMuscomicObject]{CreateMuscomicObject()}}, \code{\link[=mergeCounts]{mergeCounts()}}
}
