% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaCalling.R
\name{aggregateCounts}
\alias{aggregateCounts}
\title{Aggregate counts for \code{muscadet} objects}
\usage{
aggregateCounts(x, reference, nor.het = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{A \code{\link[=muscadet-class]{muscadet}} object containing sample data
(\code{muscadet}). This object must include clustering assignments in the
\code{cnacalling$clusters} slot.}

\item{reference}{A \code{\link[=muscadet-class]{muscadet}} object containing reference
data (\code{muscadet}).}

\item{nor.het}{A logical value to specify if normal reference allele counts
are modified to: total normal depth counts divided by 2, to force these
positions to be heterozygous in the normal reference in allelic data (e.g.
when reference cells are from a different individual but
individual-specific heterozygous positions based on matched bulk sequencing
data are used and thereby assumed to be heterozygous) before combining coverage and
allelic data. Default is \code{TRUE}.}

\item{quiet}{Logical. If \code{TRUE}, suppresses informative messages during
execution. Default is \code{FALSE}.}
}
\value{
A modified \code{\link[=muscadet-class]{muscadet}} object corresponding to the \code{x}
\code{\link[=muscadet-class]{muscadet}} object, with updated \code{cnacalling} slot containing:
\itemize{
\item \code{allelic.counts}: Processed allelic counts on variant positions, for all omics.
\item \code{coverage.counts}: Processed coverage counts merged with the reference.
\item \code{combined.counts}: Combined data for allelic and coverage counts.
}
Abbreviations:
\itemize{
\item RD = Reference allele read depth
\item AD = Alternative allele read depth
\item DP = Total read depth
\item TUM = tumor sample
\item NOR = normal reference
}
}
\description{
This function combines allelic (counts at variant positions, either common
SNPs or individual-specific heterozygous positions) and coverage counts
(counts on features) from all omics per cluster for both sample and
reference. The resulting merged data is stored in the \code{cnacalling} slot of
the sample \code{\link[=muscadet-class]{muscadet}} object.
}
\examples{
# Load example muscadet object
# data("exdata_muscadet")
# data("exdata_muscadet_ref")

# Aggregate counts from all omics from both sample and reference
exdata_muscadet <- aggregateCounts(exdata_muscadet, exdata_muscadet_ref)

}
