% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barcodes}
\alias{barcodes}
\alias{barcodes_atac_tumor}
\alias{barcodes_atac_ref}
\alias{barcodes_rna_tumor}
\alias{barcodes_rna_ref}
\title{Cell barcodes}
\format{
A character vector.
}
\usage{
barcodes_atac_tumor

barcodes_atac_ref

barcodes_rna_tumor

barcodes_rna_ref
}
\description{
Character vectors of unique cell barcodes.
}
\keyword{datasets}
