% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf.R
\name{bessel_pcf_estimation}
\alias{bessel_pcf_estimation}
\title{Estimation of Stationary Bivariate Bessel DPP via PCF}
\usage{
bessel_pcf_estimation(X, tau_init = 0.5, eta = 0.1, type = "joint")
}
\arguments{
\item{X}{A 2-dimensional marked \code{\link[spatstat]{ppp}} object.}

\item{pcftheomarginal}{A function implementing the theoretical marginal PCF.}

\item{pcftheocross}{A function implementing the theoretical cross PCF.}
}
\value{
The estimated model parameters as a list.
}
\description{
Estimation of Stationary Bivariate Bessel DPP via PCF
}
\examples{
X <- X05list[[1]]
bessel_pcf_estimation(X)

tau <- sqrt(0.5)
res <- lapply(X05list, bessel_pcf_estimation)
temp <- unlist(res)
par(mfrow = c(2, 2))
boxplot(temp[seq(2, 600, 6)], main = 'alpha1')
abline(h = 0.03)
boxplot(temp[seq(4, 600, 6)], main = 'alpha2')
abline(h = 0.03)
boxplot(temp[seq(5, 600, 6)], main = 'tau')
abline(h = tau)
boxplot(temp[seq(6, 600, 6)], main = 'alpha12')
abline(h = 0.03)

tau <- 0
res <- lapply(X0list, bessel_pcf_estimation)
temp <- unlist(res)
par(mfrow = c(2, 2))
boxplot(temp[seq(2, 600, 6)], main = 'alpha1')
abline(h = 0.03)
boxplot(temp[seq(4, 600, 6)], main = 'alpha2')
abline(h = 0.03)
boxplot(temp[seq(5, 600, 6)], main = 'tau')
abline(h = tau)
boxplot(temp[seq(6, 600, 6)], main = 'alpha12')
abline(h = 0.03)

tau <- 1
res <- lapply(X1list, bessel_pcf_estimation)
temp <- unlist(res)
par(mfrow = c(2, 2))
boxplot(temp[seq(2, 600, 6)], main = 'alpha1')
abline(h = 0.03)
boxplot(temp[seq(4, 600, 6)], main = 'alpha2')
abline(h = 0.03)
boxplot(temp[seq(5, 600, 6)], main = 'tau')
abline(h = tau)
boxplot(temp[seq(6, 600, 6)], main = 'alpha12')
abline(h = 0.03)
}
