% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_hclust}
\alias{cluster_hclust}
\title{Multi Omics Clustering with SNF integration and Hierarchical Clustering}
\usage{
cluster_hclust(
  mat_list,
  k_range = seq(2, 10, 1),
  dist_method = "euclidean",
  hclust_method = "ward.D",
  weights = rep(1, length(mat_list)),
  knn_affinity = 40,
  var_affinity = 1,
  knn_SNF = 40,
  iter_SNF = 50,
  knn_umap = 20,
  quiet = FALSE
)
}
\arguments{
\item{mat_list}{A named list of log R ratio matrices (cells x features), one
per omic layer (\code{list}).}

\item{k_range}{A numeric vector of integers (≥2) specifying the cluster
numbers (k) to extract from hierarchical clustering (\code{numeric} vector).
Default is from 2 to 10.}

\item{dist_method}{A string specifying the distance method for
\code{\link[Rfast:Dist]{Rfast::Dist()}} (e.g., \code{"euclidean"}, \code{"manhattan"}, \code{"cosine"})
(\code{character} string). Default is \code{"euclidean"}.}

\item{hclust_method}{A string specifying the hierarchical clustering linkage
method for \code{\link[fastcluster:hclust]{fastcluster::hclust()}} (e.g., \code{"ward.D"}, \code{"average"})
(\code{character} string). Default is \code{"ward.D"}.}

\item{weights}{A numeric vector of non-negative values of length equal to the
number of omic (internally normalized to sum to 1) (\code{numeric} vector). It
specifies the relatives weights of each omic for SNF with
\code{\link[=weightedSNF]{weightedSNF()}}. Omics with a weight of 0 will not contribute to
the clustering. If \code{NULL} (default), weights are uniform.}

\item{knn_affinity}{Integer specifying the number of nearest neighbors used
when building affinity matrices with \code{\link[SNFtool:affinityMatrix]{SNFtool::affinityMatrix()}}
(\code{integer}). Default is \code{40}.}

\item{var_affinity}{Numeric value for the variance parameter (Gaussian kernel
width \code{sigma}) when building affinity matrix with
\code{\link[SNFtool:affinityMatrix]{SNFtool::affinityMatrix()}} (\code{numeric}). Default is \code{1}.}

\item{knn_SNF}{Integer specifying the number of nearest neighbors used during
the Similarity Network Fusion (SNF) with \code{\link[=weightedSNF]{weightedSNF()}}
(\code{integer}). Default is \code{40}.}

\item{iter_SNF}{Integer specifying the number of iterations for SNF with
\code{\link[=weightedSNF]{weightedSNF()}} (\code{integer}). Default is \code{50}.}

\item{knn_umap}{Integer specifying the number of nearest neighbors used for
manifold approximation (UMAP) (see \code{\link[uwot:umap]{uwot::umap()}}). Default is \code{20}.}

\item{quiet}{Logical. If \code{TRUE}, suppresses informative messages during
execution. Default is \code{FALSE}.}
}
\value{
A list containing:
\describe{
\item{params}{List of parameters used for clustering (\code{list}).}
\item{SNF}{Fused similarity matrix computed with SNF (\code{matrix}).}
\item{dist}{Distance matrix derived from the SNF similarity (\code{matrix}).}
\item{hclust}{Hierarchical clustering object from \code{\link[fastcluster:hclust]{fastcluster::hclust()}} (\code{hclust}).}
\item{umap}{UMAP coordinates (\code{matrix}).}
\item{clusters}{A named list of clustering results (vectors of cluster
labels) for each value in \code{k_range} (\code{list}).}
}
}
\description{
Performs the integration of log R ratio matrices (\code{mat_list}) using
Similarity Network Fusion (SNF) followed by hierarchical clustering, on the
integrated SNF matrix, to identify clusters of cells for each specified k
number of cluster (\code{k_range}). For more than 1 omic, the integration and
clustering are performed only on common cells between omics.
}
\details{
The function calculates pairwise distances between cells within each omic dataset
using the specified \code{dist_method} (using only common cells between omics).

It constructs affinity matrices based on these distances, applies SNF to
generate a fused similarity matrix.

\code{weights} can be assigned to each omic dataset to prioritize certain data types
over others, allowing users to tailor the analysis based on the
characteristics and importance of each dataset.

It then performs a hierarchical clustering using the specified
\code{hclust_method} to assign clusters to each common cells.

Results are given as cluster assignments for each number of cluster specified
by \code{k_range}.
}
\examples{
# Load example muscadet object
data(muscadet_obj)

# Format input
# transpose matrices to: cells x features matrices
mat_list <- lapply(muscadet::matLogRatio(muscadet_obj), t)

# Run integration & clustering
result <- cluster_hclust(mat_list, k_range = 2:4)

# View results
lapply(result$clusters, table)

}
\seealso{
Similarity Network Fusion: \code{\link[=weightedSNF]{weightedSNF()}}.
}
