% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_seurat}
\alias{cluster_seurat}
\title{Multi Omics Clustering using Seurat Multi Modal Graph-based Clustering}
\usage{
cluster_seurat(
  mat_list,
  res_range = seq(0.1, 0.5, 0.1),
  dims_list = rep(list(1:8), length(mat_list)),
  algorithm = 4,
  leiden_method = "igraph",
  knn_seurat = 20,
  knn_range_seurat = 200,
  max_dim = 200,
  random.seed = 1,
  quiet = FALSE
)
}
\arguments{
\item{mat_list}{A named list of log R ratio matrices (cells x features), one
per omic layer (\code{list}).}

\item{res_range}{A numeric non-negative vector specifying the resolution
values to use for \code{\link[Seurat:FindClusters]{Seurat::FindClusters()}} (\code{numeric} vector). Default is
\code{c(0.1, 0.2, 0.3, 0.4, 0.5)}.}

\item{dims_list}{A list of vectors of PC dimensions to use for each omic
(\code{list}). Must match the length of \code{mat_list} (e.g., list(1:8) for 1 omic ;
list(1:8, 1:8) for 2 omics). Default is the first 8 dimensions for each
provided omic.}

\item{algorithm}{Integer specifying the algorithm for modularity optimization
by \code{\link[Seurat:FindClusters]{Seurat::FindClusters()}} (\code{1} = original Louvain algorithm; \code{2} =
Louvain algorithm with multilevel refinement; \code{3} = SLM algorithm; \code{4} =
Leiden algorithm). Default is \code{4} for Leiden algorithm as recommended,
see \code{\link[=cluster_seurat]{cluster_seurat()}} Details section.}

\item{leiden_method}{Character string to choose from which package for
running leiden algorithm, either leidenbase ("leidenbase") or igraph
("igraph") packages (\code{character}). Default is "igraph".}

\item{knn_seurat}{Integer specifying the number of nearest neighbors used for
graph construction with \code{\link[Seurat:Seurat]{Seurat::Seurat()}} functions \code{\link[Seurat:FindNeighbors]{Seurat::FindNeighbors()}}
(\code{k.param}) or \code{\link[Seurat:FindMultiModalNeighbors]{Seurat::FindMultiModalNeighbors()}} (\code{k.nn}) (\code{integer}).
Default is \code{20}.}

\item{knn_range_seurat}{Integer specifying the approximate number of nearest
neighbors to compute for \code{\link[Seurat:FindMultiModalNeighbors]{Seurat::FindMultiModalNeighbors()}} (\code{knn.range})
(\code{integer}). Default is \code{200}.}

\item{max_dim}{Integer specifying the maximum number of principal components
to be used for PCA computation with \code{\link[stats:prcomp]{stats::prcomp()}} (\code{integer}). Default
is \code{200}.}

\item{random.seed}{Integer specifying the seed of the random number
generator, must be greater than 0 for Leiden algorithm. (\code{integer}).
Default is \code{1}.}

\item{quiet}{Logical. If \code{TRUE}, suppresses informative messages during
execution. Default is \code{FALSE}.}
}
\value{
A list containing:
\describe{
\item{params}{List of parameters used for clustering (\code{list}).}
\item{pcs}{List of principal components summaries for each omic (\code{list} of
\code{\link[stats:summary.prcomp]{summary.prcomp}}).}
\item{nn}{Nearest neighbors object (\code{\link[SeuratObject:Neighbor-class]{Neighbor}}).}
\item{graph}{Shared nearest neighbors graph (\code{\link[SeuratObject:Graph-class]{Graph}}).}
\item{dist}{Distance matrix derived from the graph (\code{matrix}).}
\item{umap}{UMAP coordinates (\code{matrix}).}
\item{clusters}{A named list of clustering results (vectors of cluster
labels) for each value in \code{res_range} (\code{list}).}
}
}
\description{
Performs graph-based clustering of cells using
\code{\link[Seurat:Seurat-package]{Seurat}}, based on one or two log R ratio matrices
(\code{mat_list}), including shared nearest neighbors (SNN) graph construction on
selected dimensions from PCA (\code{dims_list}), to identify clusters of cells for
each specified resolution (\code{res_range}).
\itemize{
\item For two omics: multimodal integration is performed using
\code{\link[Seurat:FindMultiModalNeighbors]{Seurat::FindMultiModalNeighbors()}} (weighted shared nearest neighbors graph).
Only common cells between omics are used.
\item For a single omic: \code{\link[Seurat:FindNeighbors]{Seurat::FindNeighbors()}} (shared nearest neighbors
graph) is used.
}
}
\details{
The Leiden algorithm (\code{algorithm = 4}) is recommended based on published work and best-practice
guidelines:
\itemize{
\item Traag, V.A., Waltman, L. & van Eck, N.J. \emph{From Louvain to Leiden: guaranteeing well-connected communities.} Sci Rep 9, 5233 (2019).
\url{https://doi.org/10.1038/s41598-019-41695-z>}
\item Heumos, L., Schaar, A.C., Lance, C. et al. \emph{Best practices for single-cell
analysis across modalities.} Nat Rev Genet (2023).
\url{https://doi.org/10.1038/s41576-023-00586-w}
\url{https://www.sc-best-practices.org/cellular_structure/clustering.html}
}

Explanation about igraph being the default package for Leiden implementation
(\code{leiden_method = "igraph"}) here: https://github.com/satijalab/seurat/issues/9800
}
\examples{
\dontrun{
# Load example muscadet object
# data("exdata_muscadet")

# Format input
# transpose matrices to: cells x features matrices
mat_list <- matLogRatio(exdata_muscadet)

# Run integration & clustering
result <- cluster_seurat(
    mat_list,
    res_range = c(0.1, 0.3, 0.5),
    knn_seurat = 10,
    knn_range_seurat = 30
)

# View results
lapply(result$clusters, table)
}

}
\seealso{
\href{https://satijalab.org/seurat/articles/weighted_nearest_neighbor_analysis}{Weighted Nearest Neighbor Analysis Vignette from Seurat}
}
