% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLogRatio.R
\name{computeLogRatio}
\alias{computeLogRatio}
\title{Compute log R ratios}
\usage{
computeLogRatio(
  x,
  reference,
  omic,
  method = NULL,
  new.label.features = NULL,
  quiet = FALSE,
  all_steps = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{\link[=muscadet-class]{muscadet}} object containing sample data
(\code{muscadet}).}

\item{reference}{Another \code{\link[=muscadet-class]{muscadet}} object containing
reference data (\code{muscadet}).}

\item{omic}{Name of the omic to apply this function (\code{character} string).}

\item{method}{Method to apply to the selected omic (\code{character} string).
Supported methods are "ATAC" and "RNA":
\itemize{
\item "ATAC" method calls for \code{\link[=computeLogRatioATAC]{computeLogRatioATAC()}} function
\item "RNA" method calls for \code{\link[=computeLogRatioRNA]{computeLogRatioRNA()}} function
}

If \code{NULL} and if the omic type is either "ATAC" or "RNA", the corresponding
method will be applied.}

\item{new.label.features}{New label for features (\code{character} string). If
\code{NULL}, the label remains unchanged when using the "RNA" method and becomes
"windows of peaks" when using the "ATAC" method.}

\item{quiet}{Logical. If \code{TRUE}, suppresses informative messages during
execution. Default is \code{FALSE}.}

\item{all_steps}{Logical value to indicate whether matrices at each step of
the log ratio computation are kept and returned (\code{logical}). IMPORTANT: if
\code{TRUE}, the function does not return an updated
\code{\link[=muscadet-class]{muscadet}} object but a list with matrices at each step
of the log ratio computation (see Value section). Default is \code{FALSE}.}

\item{...}{
  Arguments passed on to \code{\link{computeLogRatioATAC}}, \code{\link{computeLogRatioRNA}}
  \describe{
    \item{\code{windowSize}}{Size of windows in base pairs (\code{integer} value). By default: \code{10e6} (10
Mbp).}
    \item{\code{slidingSize}}{Distance between start positions of sliding windows in base pairs
(\code{integer} value). If set to the same value as \code{windowSize}, the windows
don't overlap. By default: \code{2e6} (2 Mbp).}
    \item{\code{minReads}}{Minimum read average per window in reference cells (\code{integer}
value). By default: \code{5}.}
    \item{\code{minPeaks}}{Minimum number of peaks per window (\code{integer} value). By
default: \code{100}.}
    \item{\code{genesPerWindow}}{Number of genes per moving window (\code{integer} value). By
default: \code{101}.}
    \item{\code{refReads}}{Minimum of reads in reference cells (\code{integer} value). By
default: \code{100}.}
    \item{\code{refMeanReads}}{Minimum of average reads per reference cell (\code{integer}
value). By default: \code{0.01}.}
    \item{\code{thresh_capping}}{Threshold to cap the range of log R ratio values
(\code{numeric} value). By default: \code{3}.}
  }}
}
\value{
A \code{\link[=muscadet-class]{muscadet}} object corresponding to the sample
\code{\link[=muscadet-class]{muscadet}} object (\code{x}) containing the computed log R ratio
matrix in the \code{coverage} slot of the selected \code{omic}.

If the \code{all_steps} argument is set to \code{TRUE}, it returns a list with intermediate
matrices \code{matTumor} and \code{matRef} from every step from \code{step01} to \code{step08}
(no \code{step06} for \code{method} = "ATAC"), \code{coord} table of features coordinates
and associated data, and \code{params} list of parameters used for the function.
}
\description{
Computes log R ratios from raw count matrices within
\code{\link[=muscadet-class]{muscadet}} objects. The log R ratios values are computed
based on read counts from a sample \code{\link[=muscadet-class]{muscadet}} object
\emph{versus} read counts a the \code{reference} \code{\link[=muscadet-class]{muscadet}} object.
In the output sample \code{\link[=muscadet-class]{muscadet}} object, the newly computed
matrix of log R ratios is added for the selected \code{omic}.
}
\details{
Log R ratios computation steps are described in functions:
\itemize{
\item \code{\link[=computeLogRatioATAC]{computeLogRatioATAC()}}
\item \code{\link[=computeLogRatioRNA]{computeLogRatioRNA()}}
}
}
\examples{
# Create muscomic objects
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = exdata_mat_counts_atac_tumor,
  allele_counts = exdata_allele_counts_atac_tumor,
  features = exdata_peaks
)
rna <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = exdata_mat_counts_rna_tumor,
  allele_counts = exdata_allele_counts_rna_tumor,
  features = exdata_genes
)
atac_ref <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = exdata_mat_counts_atac_ref,
  allele_counts = exdata_allele_counts_atac_ref,
  features = exdata_peaks
)
rna_ref <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = exdata_mat_counts_rna_ref,
  allele_counts = exdata_allele_counts_rna_ref,
  features = exdata_genes
)

# Create muscadet objects
muscadet <- CreateMuscadetObject(
  omics = list(atac, rna),
  bulk.lrr = exdata_bulk_lrr,
  bulk.label = "WGS",
  genome = "hg38"
)
muscadet_ref <- CreateMuscadetObject(
  omics = list(atac_ref, rna_ref),
  genome = "hg38"
)

# compute log R ratios for ATAC
muscadet <- computeLogRatio(
  x = muscadet,
  reference = muscadet_ref,
  omic = "ATAC",
  method = "ATAC",
  minReads = 1, # low value for example subsampled datasets
  minPeaks = 1 # low value for example subsampled datasets
)

# compute log R ratios for RNA
muscadet <- computeLogRatio(
  x = muscadet,
  reference = muscadet_ref,
  omic = "RNA",
  method = "RNA",
  refReads = 2 # low value for example subsampled datasets
)

}
\seealso{
Other computeLogRatio:
\code{\link[=computeLogRatioATAC]{computeLogRatioATAC()}},
\code{\link[=computeLogRatioRNA]{computeLogRatioRNA()}}
}
\concept{computeLogRatio}
