% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLogRatio.R
\name{computeLogRatioATAC}
\alias{computeLogRatioATAC}
\title{Compute log R ratios for scATAC-seq data}
\usage{
computeLogRatioATAC(
  matTumor,
  matRef,
  peaksCoord,
  genome = "hg38",
  windowSize = 1e+07,
  slidingSize = 2e+06,
  minReads = 5,
  minPeaks = 100,
  thresh_capping = 3,
  all_steps = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{matTumor}{Raw count matrix \emph{features x cells} for tumor/sample cells
(\code{matrix} or \code{dgCMatrix}).}

\item{matRef}{Raw count matrix \emph{features x cells} for reference cells
(\code{matrix} or \code{dgCMatrix}).}

\item{peaksCoord}{Data frame of peak coordinates with columns \code{CHROM}, \code{start},
\code{end}, \code{id} (\code{data.frame}).}

\item{genome}{Reference genome name among: "hg38", "hg19" and "mm10"
(\code{character} string). By default: "hg38".}

\item{windowSize}{Size of windows in base pairs (\code{integer} value). By default: \code{10e6} (10
Mbp).}

\item{slidingSize}{Distance between start positions of sliding windows in base pairs
(\code{integer} value). If set to the same value as \code{windowSize}, the windows
don't overlap. By default: \code{2e6} (2 Mbp).}

\item{minReads}{Minimum read average per window in reference cells (\code{integer}
value). By default: \code{5}.}

\item{minPeaks}{Minimum number of peaks per window (\code{integer} value). By
default: \code{100}.}

\item{thresh_capping}{Threshold to cap the range of log R ratio values
(\code{numeric} value). By default: \code{3}.}

\item{all_steps}{\code{TRUE} or \code{FALSE} (\code{logical}). Whether to keep intermediate
result from every step in the final object. By default: \code{FALSE}.}

\item{quiet}{Logical. If \code{TRUE}, suppresses informative messages during
execution. Default is \code{FALSE}.}
}
\value{
If \code{all_steps} is set to \code{FALSE}, a list containing:
\describe{
\item{\code{matTumor}}{Matrix of log R ratio values \emph{cells x features} for the
tumor/sample cells (\code{matrix}).}
\item{\code{matRef}}{Matrix of log R ratio values \emph{cells x features} for the
reference cells (\code{matrix}).}
\item{\code{params}}{List of parameters set for the \code{windowSize}, \code{slidingSize},
\code{minReads}, \code{minPeaks} and \code{thresh_capping} arguments (\code{list}).}
\item{\code{coord}}{Data frame of coordinates for windows of peaks and associated
data along the different steps (\code{data.frame}).
Columns:
\itemize{
\item \code{CHROM}, \code{start}, \code{end}, \code{width}, \code{id}: coordinates and unique identifier of
windows (depends on \code{windowSize} and \code{slidingSize} arguments).
\item \code{nPeaks}: number of peaks per window.
\item \code{sumReads.tum/ref}: sum of read counts for all cells in tumor or reference
cells.
\item \code{meanReads.tum/ref}: mean of read counts per cells for tumor or reference
cells.
\item \code{sdReads.tum/ref}: standard deviation of read counts per cells for tumor
or reference cells.
\item \code{keep}: logical, \code{TRUE} for windows to keep after filtering based on
coverage (depends on \code{minPeaks} and \code{meanReads} arguments).
\item \code{meanReads/sdReads.norm.tum/ref}: mean/sd of normalized counts per million
for tumor/reference cells.
\item \code{meanLRR/sdReads.raw.tum/ref}: mean/sd of raw log R ratio (LRR) for
tumor/reference cells.
\item \code{meanLRR/sdLRR.cap.tum/ref}: mean/sd of capped log R ratio (LRR) for
tumor/reference cells (depends on \code{thresh_capping} argument).
\item \code{meanLRR/sdLRR.cent.tum/ref}: mean/sd of centered log R ratio (LRR) for
tumor/reference cells.
\item \code{meanLRR/sdLRR.corr.tum/ref}: mean/sd of final log R ratio (LRR) corrected
by reference variability for tumor/reference cells.}
}
If \code{all_steps} is set to \code{TRUE}, the previously described list can be found for
each step in a different list element.
}
}
\description{
Compute log R ratios from raw count matrices with method specifically adapted
for scATAC-seq data. The counts per peaks are grouped into counts per windows
of peaks, thereby features become windows of peaks instead of peaks.
}
\details{
The raw count matrix is transformed into log R ratios through the following
steps:
\itemize{
\item Group peaks in windows (\code{windowSize} and \code{slidingSize} arguments)
\item Filtering on coverage (\code{minPeaks} and \code{meanReads} arguments)
\item Normalization for sequencing depth
\item Log transformation and normalization by reference data: log R ratio
\item Capping the range of values (\code{thresh_capping} argument)
\item Centering of cells
\item Correcting by reference variability
}
}
\examples{
# Create muscomic objects
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = exdata_mat_counts_atac_tumor,
  features = exdata_peaks
)
rna <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = exdata_mat_counts_rna_tumor,
  features = exdata_genes
)
atac_ref <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = exdata_mat_counts_atac_ref,
  features = exdata_peaks
)
rna_ref <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = exdata_mat_counts_rna_ref,
  features = exdata_genes
)

# Create muscadet objects
muscadet <- CreateMuscadetObject(
  omics = list(atac, rna),
  bulk.lrr = exdata_bulk_lrr,
  bulk.label = "WGS",
  genome = "hg38"
)
muscadet_ref <- CreateMuscadetObject(
  omics = list(atac_ref, rna_ref),
  genome = "hg38"
)

# Compute log R ratio for ATAC
obj_atac <- computeLogRatioATAC(
  matTumor = matCounts(muscadet)$ATAC,
  matRef = matCounts(muscadet_ref)$ATAC,
  peaksCoord = coordFeatures(muscadet)$ATAC,
  genome = muscadet$genome,
  minReads = 1, # low value for example subsampled datasets
  minPeaks = 1 # low value for example subsampled datasets
)
table(obj_atac$coord$keep)

# With results form every step when `all_steps = TRUE`
obj_atac_all <- computeLogRatioATAC(
  matTumor = matCounts(muscadet)$ATAC,
  matRef = matCounts(muscadet_ref)$ATAC,
  peaksCoord = coordFeatures(muscadet)$ATAC,
  genome = muscadet$genome,
  minReads = 1, # low value for example subsampled datasets
  minPeaks = 1, # low value for example subsampled datasets
  all_steps = TRUE
)
names(obj_atac_all)
table(obj_atac_all$coord$keep)
nrow(obj_atac_all$step08$matTumor)

}
\seealso{
Other computeLogRatio:
\code{\link[=computeLogRatio]{computeLogRatio()}},
\code{\link[=computeLogRatioRNA]{computeLogRatioRNA()}}
}
\concept{computeLogRatio}
