% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddetmodel.R
\name{ddetmodel.direct}
\alias{ddetmodel.direct}
\title{Density of determinantal point process model via direct calculation}
\usage{
ddetmodel.direct(X, model, trunc = 0.99, log = TRUE, ...)
}
\arguments{
\item{X}{An object of class \link[spatstat.geom:ppp]{spatstat.geom::ppp} specifying a planar point
pattern.}

\item{model}{An object of class \link[spatstat.core:detpointprocfamilyfun]{spatstat.core::detpointprocfamilyfun}
specifying the model to be used.}

\item{trunc}{A numeric value specifying how the model truncation is
preformed. Defaults to \code{0.99}. See details of
\code{\link[spatstat.core:simulate.dppm]{spatstat.core::simulate.detpointprocfamily()}}.}

\item{log}{A boolean value specifying whether to compute the logarithm.
Defaults to \code{TRUE}.}

\item{...}{Extra parameters to be passed on to next methods. Ignored.}
}
\description{
Evaluates the density of a determinantal point process model at a given point
pattern via direct calculation (i.e. not using FFT, etc.).
}
\keyword{internal}
