% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstimateBessel}
\alias{EstimateBessel}
\title{Stationary Bivariate Bessel DPP Estimator}
\usage{
EstimateBessel(X, labels, lb, ub, amplitude1 = NA_real_,
  amplitude2 = NA_real_, amplitude12 = NA_real_, alpha1 = NA_real_,
  alpha2 = NA_real_, alpha12 = NA_real_)
}
\arguments{
\item{X}{A matrix of size n x (d+1) storing the points in R^d and their label in last column.}
}
\value{
A vector with the estimated model parameters.
}
\description{
This function estimates the parameters of a stationary bivariate Gaussian DPP from a set of observed points and labels.
}
\examples{
library(spatstat)
# Simulate some data
m <- dppGauss(lambda = 100, alpha = 0.05, d = 2)
X1 <- simulate(m)
fit1 <- dppm(X1, dppGauss, method = "palm")$fitted$fixedpar
X2 <- stats::simulate(m)
fit2 <- dppm(X2, dppGauss, method = "palm")$fitted$fixedpar
X1 <- cbind(X1$x, X1$y, rep(1, X1$n))
X2 <- cbind(X2$x, X2$y, rep(2, X2$n))
X <- rbind(X1, X2)

# Run parameter estimation
params <- Estimate(X, fit1$lambda, fit2$lambda, fit1$alpha, fit2$alpha)

# Verify parameters were recovered
params
}
