% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exdata_allele_counts}
\alias{exdata_allele_counts}
\alias{exdata_allele_counts_atac_tumor}
\alias{exdata_allele_counts_atac_ref}
\alias{exdata_allele_counts_rna_tumor}
\alias{exdata_allele_counts_rna_ref}
\title{Example data: Allele counts at variant positions}
\format{
A data frame with columns based on the \href{https://en.wikipedia.org/wiki/Variant_Call_Format}{Variant Call Format (VCF)} columns. It
contains the following columns:
\describe{
\item{\code{cell}}{Barcodes of cells.}
\item{\code{id}}{Variant unique identifier defined as
CHROM_POS_REF_ALT, e.g. "3_3126620_T_G".}
\item{\code{CHROM}}{Chromosome names, e.g. "3", "X".}
\item{\code{POS}}{Position of the variant (1-base positions).}
\item{\code{REF}}{Reference allele base, "A" "C" "G" or "T".}
\item{\code{ALT}}{Alternative allele base, "A" "C" "G" or "T".}
\item{\code{RD}}{Reference allele depth/count.}
\item{\code{AD}}{Alternative allele depth/count.}
\item{\code{DP}}{Total depth/count.}
\item{\code{GT}}{Genotype: "0/1" or "1/0" if unphased; "0|1" or "1|0" if phased.}
}
}
\usage{
exdata_allele_counts_atac_tumor

exdata_allele_counts_atac_ref

exdata_allele_counts_rna_tumor

exdata_allele_counts_rna_ref
}
\description{
Data frames of allele counts at variant positions per cell. Variant positions
can be either common single nucleotide polymorphisms (SNPs) positions or
individual-specific heterozygous positions retrieved using bulk sequencing.
Only chromosomes 3, 4 and 8 are present in the \code{exdata} example dataset.
}
\seealso{
\itemize{
\item \code{\link[=process_allele]{process_allele()}}
}
}
\keyword{datasets}
