% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exdata_bulk_lrr}
\alias{exdata_bulk_lrr}
\title{Example data: Log R ratio from bulk sequencing data}
\format{
A data frame with the following columns:
\describe{
\item{\code{CHROM}}{Chromosome names, e.g. "3", "X".}
\item{\code{start}}{Start position of the segment.}
\item{\code{end}}{End position of the segment.}
\item{\code{lrr}}{Log R ratio of the segment ("cnlr.median" column from
\code{\link[facets:fitcncf]{facets::fitcncf()}} \verb{$cncf} data frame).}
}
}
\usage{
exdata_bulk_lrr
}
\description{
Data frame containing log R ratio values per genomic segments from bulk
sequencing data. Only chromosomes 3, 4 and 8 are present in the \code{exdata}
example dataset.
}
\note{
Data obtained from whole genome sequencing (WGS) after using
\code{\link[facets:fitcncf]{facets::fitcncf()}} from \code{\link[facets:facets-package]{facets}} "Cellular
Fraction and Copy Numbers from Tumor Sequencing" version \verb{0.6.2}: \verb{$cncf}
data frame columns \code{chrom}, \code{start}, \code{end}, and \code{cnlr.median}.
}
\references{
\describe{
\item{\code{\link[facets:facets-package]{facets}} package}{Shen R, Seshan VE. FACETS:
allele-specific copy number and clonal heterogeneity analysis tool for
high-throughput DNA sequencing. Nucleic Acids Res. 2016 Sep 19;44(16):e131.
doi: \href{https://www.doi.org/10.1093/nar/gkw520}{10.1093/nar/gkw520}. PMID:
27270079; PMCID: PMC5027494.}
}
}
\keyword{datasets}
