% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exdata_features}
\alias{exdata_features}
\alias{exdata_genes}
\alias{exdata_peaks}
\title{Example data: Feature coordinates}
\format{
A data frame with the following columns:
\describe{
\item{\code{CHROM}}{Chromosome names, e.g. "3", "X".}
\item{\code{start}}{Start positions.}
\item{\code{end}}{End positions.}
\item{\code{id}}{Unique identifiers, e.g. gene name "CDH1" or peak identifier
CHROM_start_end "1_1600338_1600838".}
}
}
\usage{
exdata_genes

exdata_peaks
}
\description{
Data frames of features (peaks, genes...) coordinates on genome. Only
chromosomes 3, 4 and 8 are present in the \code{exdata} example dataset.
}
\keyword{datasets}
