% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exdata_mat_counts}
\alias{exdata_mat_counts}
\alias{exdata_mat_counts_atac_tumor}
\alias{exdata_mat_counts_atac_ref}
\alias{exdata_mat_counts_rna_tumor}
\alias{exdata_mat_counts_rna_ref}
\title{Example data: Matrices of raw counts}
\format{
A \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} of numeric values with the following
dimensions:
\describe{
\item{\code{rows}}{Features (peaks, genes, ...).}
\item{\code{columns}}{Cell barcodes.}
}
}
\usage{
exdata_mat_counts_atac_tumor

exdata_mat_counts_atac_ref

exdata_mat_counts_rna_tumor

exdata_mat_counts_rna_ref
}
\description{
Matrices of raw read counts \emph{features x cells} in
\code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} format.
}
\keyword{datasets}
