% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exdata_muscadet}
\alias{exdata_muscadet}
\alias{exdata_muscadet_ref}
\title{Example data: muscadet objects}
\format{
\code{\link[=muscadet-class]{muscadet}} objects with the following slots:
\describe{
\item{\code{omics}}{List of \code{\link[=muscomic-class]{muscomic}} objects, one per
single-cell omic (\code{list}).}
\item{\code{bulk.data}}{List of data from paired bulk sequencing (\code{list}).}
\item{\code{clustering}}{List of data associated with the clustering of cells
based on coverage log R ratio values (\code{list}).}
\item{\code{cnacalling}}{List of data associated with the calling of copy number
alterations (CNAs) (\code{list}).}
\item{\code{genome}}{Reference genome name among: "hg38", "hg19" and "mm10" (\code{character}).}
}
}
\usage{
exdata_muscadet

exdata_muscadet_ref
}
\description{
\code{\link[=muscadet-class]{muscadet}} objects, containing two
single-cell omic datasets: scATAC-seq and scRNA-seq.
\itemize{
\item \code{exdata_muscadet} with tumor cells data: sample cells
\item \code{exdata_muscadet_ref} with normal cells data: reference cells
}
}
\keyword{datasets}
