% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features}
\alias{features}
\alias{genes}
\alias{peaks}
\title{Feature coordinates}
\format{
A data frame with the following columns:
\describe{
\item{\code{CHROM}}{Chromosome in chracter format, e.g. "15", "X" (\code{character}).}
\item{\code{start}}{Start position of the feature (\code{integer}).}
\item{\code{end}}{End position of the feature (\code{character}).}
\item{\code{id}}{Unique identifier, e.g. gene name "CDH1" or peak identifier
CHROM_start_end "1_1600338_1600838" (\code{character}).}
}
}
\usage{
genes

peaks
}
\description{
Data frames of features (peaks, genes...) coordinates on genome.
}
\keyword{datasets}
