% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{fit_via_mle}
\alias{fit_via_mle}
\title{Estimation of Two-Mark Planar DPPs via Maximum Likelihood}
\usage{
fit_via_mle(
  X,
  initial_guess = NULL,
  fixed_marginal_parameters = FALSE,
  model = "gauss",
  optimizer = "bobyqa",
  num_threads = 1,
  N = 512,
  verbose_level = 0
)
}
\arguments{
\item{X}{An object of class \link[spatstat.geom:ppp]{spatstat.geom::ppp} specifying an observed
planar determinantal point process.}

\item{initial_guess}{A numeric vector specifying an initial guess for the
model parameters that maximize the likelihood. Defaults to \code{NULL}, which
initializes at \eqn{0.5} all parameters after suitable transformation into
\eqn{[0, 1]}. If provided, expected order is \code{c(alpha1, alpha2, alpha12, tau)}.}

\item{fixed_marginal_parameters}{A boolean value specifying whether the
marginal parameters should be estimated separately using the marginal
likelihood and then fixed to further estimate the cross parameters.
Defaults to \code{FALSE}.}

\item{model}{A string specifying a DPP model. Choices are \code{"gauss"} or
\code{"bessel"}. Defaults to \code{"gauss"}.}

\item{optimizer}{A string specifying one of the available derivative-free
optimizers in NLOpt. Defaults to \code{"bobyqa"}.}

\item{num_threads}{An integer value specifying the number of thread to run
on. Defaults to \code{1L}.}

\item{N}{An integer value specifying the maximum truncation index for Fourier
transform. Defaults to \code{512L}.}

\item{verbose_level}{An integer value specifying the display information
during optimization. Choose \code{0L} for no information, \code{1L} for global
information at likelihood setup or \code{2L} for detailed information at each
function evaluation. Defaults to `0L.}
}
\value{
A list with two components:
\itemize{
\item \code{par} (optimized model parameters);
\item \code{value} (\verb{-2 logLik} at the maximum likelihood).
}
}
\description{
Estimation of Two-Mark Planar DPPs via Maximum Likelihood
}
\examples{
opt <- fit_via_mle(sim_gauss5[[1]])
}
