% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_feasible_repulsion_rate}
\alias{get_feasible_repulsion_rate}
\title{Feasible range for marginal repulsion rates}
\usage{
get_feasible_repulsion_rate(rho, d = 2, model = c("Gauss", "Bessel"))
}
\arguments{
\item{rho}{A numeric value specifying the intensity of the DPP.}

\item{d}{An integer value specifying the dimension of the DPP. Defaults to \code{2L}.}

\item{model}{A string specifying the model to be used. Choices are \code{"Gauss"} or \code{"Bessel"}. Defaults to \code{"Gauss"}.}
}
\value{
Displays the interval to the console but outputs \code{NULL}.
}
\description{
For a given intensity, dimension and model, provides bounds for the repulsion
rate.
}
\examples{
get_feasible_repulsion_rate(rho = 100)
}
