% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_feasible_space}
\alias{get_feasible_space}
\title{Feasible range for cross parameters}
\usage{
get_feasible_space(
  rho1,
  rho2,
  alpha1,
  alpha2,
  d = 2,
  model = c("Gauss", "Bessel")
)
}
\arguments{
\item{rho1}{A numeric value specifying the intensity of the first marginal
DPP.}

\item{rho2}{A numeric value specifying the intensity of the second marginal
DPP.}

\item{alpha1}{A numeric value specifying the repulsion rate of the first
marginal DPP.}

\item{alpha2}{A numeric value specifying the repulsion rate of the second
marginal DPP.}

\item{d}{An integer value specifying the dimension of the DPP. Defaults to
\code{2L}.}

\item{model}{A string specifying the model to be used. Choices are \code{"Gauss"}
or \code{"Bessel"}. Defaults to \code{"Gauss"}.}
}
\value{
Displays the intervals to the console but outputs \code{NULL}.
}
\description{
For a given set of marginal parameters (intensities and repulsion rates),
provides bounds for the cross repulsion rate and the correlation between
marks.
}
\examples{
get_feasible_space(rho1 = 100, rho2 = 100, alpha1 = 0.03, alpha2 = 0.03)
}
