% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLogRatio.R
\name{getLogRatioBulk}
\alias{getLogRatioBulk}
\title{Retrieve log R ratio from Bulk data on single-cell features (internal)}
\usage{
getLogRatioBulk(x, bulk.lrr)
}
\arguments{
\item{x}{A \code{muscomic} object containing the single-cell omic data, which
includes the feature coordinates (\code{muscomic}).}

\item{bulk.lrr}{A data frame containing log R ratio per genomic segments from
bulk sequencing data (\code{data.frame}). One row per segment and 4 columns
ordered as followed: chromosome (\code{integer}), start position (\code{integer}),
end position (\code{integer}), and Log R ratio value (\code{numeric}).}
}
\value{
A data frame of single-cell omic features, with columns: \code{CHR}, \code{start},
\code{end}, and \code{bulk.lrr}, where \code{bulk.lrr} corresponds to the log R ratio values
retrieved from bulk data matching the feature coordinates.
}
\description{
This internal function assigns the log R ratio (LRR) values from bulk data
segments to single-cell omic features. It matches the bulk data segments to
corresponding genomic features from the single-cell omic data and returns a
table of single-cell features with the corresponding bulk LRR values.
}
\examples{
# Create a muscomic object
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_tumor,
  allele_counts = allele_counts_atac_tumor,
  features = peaks
)
# or use a muscomic object inside a muscadet object
atac <- slot(muscadet_obj, "omics")[["ATAC"]]

# Load bulk Log R ratio data frame
data(bulk_lrr)
head(bulk_lrr)
# or use the one inside a muscadet object
head(slot(muscadet_obj, "bulk.data")[["log.ratio"]])

features_bulk_lrr <- getLogRatioBulk(
  x = atac,
  bulk.lrr = bulk_lrr # or slot(muscadet_obj, "bulk.data")[["log.ratio"]]
)
head(features_bulk_lrr)

}
\seealso{
example data: \code{\link{bulk_lrr}}
functions: \code{\link{muscomic}}, \code{\link{muscadet}}
}
