% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{imputeClusters}
\alias{imputeClusters}
\title{Impute Cluster Assignments for Missing Cells by Similarity}
\usage{
imputeClusters(mat_list, clusters, knn_imp = 10)
}
\arguments{
\item{mat_list}{A named list of log ratio \emph{cells x features} matrices where each matrix
corresponds to a single omic dataset (\code{list}). Rows are cells, and columns are
features.}

\item{clusters}{A named list for each \code{k} values (number of clusters) where
each element is a named vector of cluster assignments for cells (\code{list}).}

\item{knn_imp}{Integer specifying the number of nearest neighbors cells to
use for imputation (\code{integer}).}
}
\value{
A named list for each \code{k} values (number of clusters) where each
element is a named vector of imputed cluster assignments for cells.
}
\description{
This function imputes cluster assignments for cells missing in some omics
by leveraging nearest neighbor cells in other omic matrices.
}
\details{
The function operates in the following steps:
\enumerate{
\item Identifies cells missing in specific matrices.
\item Finds the k-nearest neighbors for missing cells in matrices where they are
present.
\item Imputes cluster assignments for missing cells based on the clusters
assigned to their neighbors.
\item Resolves ties (two major clusters found among neighbors) by selecting the
one of the first nearest neighbor.
}

The imputation is performed separately for each omic dataset, and the results
are aggregated to provide final cluster assignments.
}
\examples{
# Create matrices with some cells missing in one or the other
mat1 <- matrix(runif(100), nrow = 20)
mat2 <- matrix(runif(100), nrow = 20)
rownames(mat1) <- paste0("Cell", 1:20)
rownames(mat2) <- paste0("Cell", c(1:5, 11:25))
mat_list <- list(ATAC = mat1, RNA = mat2)

# Create cluster assignments for common cells
clusters <- list(
  `2` = setNames(sample(1:2, 15, replace = TRUE), intersect(rownames(mat1), rownames(mat2))),
  `3` = setNames(sample(1:3, 15, replace = TRUE), intersect(rownames(mat1), rownames(mat2))),
  `4` = setNames(sample(1:4, 15, replace = TRUE), intersect(rownames(mat1), rownames(mat2)))
)

# Impute cluster assignments for missing cells
imputed_clusters <- imputeClusters(mat_list, clusters, knn_imp = 5)

# View the imputed cluster assignments
print(imputed_clusters)

}
