% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeAllelicSparse}
\alias{makeAllelicSparse}
\title{Construct Sparse Matrices for Allelic Data}
\usage{
makeAllelicSparse(allele_counts)
}
\arguments{
\item{allele_counts}{A data.frame containing at least the columns:
\code{cell}, \code{CHROM}, \code{POS}, \code{REF}, \code{ALT},
\code{RD}, \code{AD}. Each row represents a single variant in a
specific cell.}
}
\value{
A list with the following elements:
\describe{
\item{mat.RD}{A \code{Matrix::dgCMatrix} sparse matrix of reference depths.}
\item{mat.AD}{A \code{Matrix::dgCMatrix} sparse matrix of alternate depths.}
\item{coord.vars}{A data.frame of variant metadata with columns:
\code{id}, \code{CHROM}, \code{POS}, \code{REF}, \code{ALT}.}
}
}
\description{
This function converts a long-format allele counts table into two sparse
matrices: one containing reference allele depths (RD) and one containing
alternate allele depths (AD). Each matrix has cells as rows and variant
positions as columns. Variant coordinates and metadata (id, CHROM, POS, REF,
ALT) is also returned in a structured data frame in the list of outputs.
}
\examples{
allele_counts <- data.frame(
  cell = c("cell1","cell1","cell2"),
  CHROM = c("1","1","1"),
  POS = c(101,102,101),
  REF = c("A","G","A"),
  ALT = c("C","T","C"),
  RD = c(10, 0, 4),
  AD = c(3, 5, 0)
)
out <- makeAllelicSparse(allele_counts)
str(out)
# access cells names
rownames(out$mat.RD)
# access variants ids
colnames(out$mat.RD) # or
out$coord.vars$id

}
