% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_counts}
\alias{mat_counts}
\alias{mat_counts_atac_tumor}
\alias{mat_counts_atac_ref}
\alias{mat_counts_rna_tumor}
\alias{mat_counts_rna_ref}
\title{Example data: Matrices of raw counts}
\format{
A \code{dgCMatrix} (\code{\link{dgCMatrix-class}}) of numeric values with the following dimensions:
\describe{
\item{\code{rows}}{Features (peaks, genes).}
\item{\code{columns}}{Cell barcodes.}
}
}
\usage{
mat_counts_atac_tumor

mat_counts_atac_ref

mat_counts_rna_tumor

mat_counts_rna_ref
}
\description{
Matrices of raw read counts \emph{features x cells} in \code{dgCMatrix} format.
}
\keyword{datasets}
