% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Estimation of multi-mark DPP via Maximum Likelihood}
\usage{
mle(
  data,
  model = "gauss",
  optimizer = "bobyqa",
  num_threads = 1,
  N = 512,
  use_verbose = FALSE
)
}
\arguments{
\item{data}{A ppp.}

\item{model}{A DPP model. For now either gauss or bessel.}

\item{optimizer}{Any derivative-free optimizer in NlOpt.}

\item{num_threads}{Number of threas to run on.}

\item{N}{Maximum...}

\item{use_verbose}{Show information along the way.}
}
\value{

}
\description{
Estimation of multi-mark DPP via Maximum Likelihood
}
