% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{[.muscadet}
\alias{[.muscadet}
\alias{muscadet-access}
\alias{[.muscomic}
\alias{$.muscadet}
\alias{$.muscomic}
\alias{[<-.muscadet}
\alias{[<-.muscomic}
\alias{$<-.muscadet}
\alias{$<-.muscomic}
\title{Access and assignment methods for \code{muscadet} objects}
\usage{
\method{[}{muscadet}(x, i, ...)

\method{[}{muscomic}(x, i, ...)

\method{$}{muscadet}(x, name)

\method{$}{muscomic}(x, name)

\method{[}{muscadet}(x, i) <- value

\method{[}{muscomic}(x, i) <- value

\method{$}{muscadet}(x, i) <- value

\method{$}{muscomic}(x, i) <- value
}
\arguments{
\item{x}{A \code{\link{muscadet}} or \code{\link{muscomic}} object.}

\item{i}{The name of the slot (or omic).}

\item{...}{Other arguments (ignored).}

\item{name}{The name of the slot (or omic).}

\item{value}{The new value to assign.}
}
\value{
The selected slot or the omic dataset (\code{\link{muscomic}} object)
for \code{muscadet} objects. The selected slot for \code{muscomic} objects.

The updated \code{\link{muscadet}} or \code{\link{muscomic}} object.
}
\description{
Simplified access to omic datasets and slots inside
\code{\link{muscadet}} objects.

Assign new data in a \code{\link{muscadet}} or
\code{\link{muscomic}} object. For \code{muscadet} objects, the omic datasets in
the \code{omics} slot can be directly reassigned.
}
\examples{
# Load example muscadet object
data(muscadet_obj)
# Access to muscadet omics or slots
muscadet_obj["ATAC"]
muscadet_obj["genome"]

# Load example muscadet object
data(muscadet_obj)
# Access to muscomic slots
muscadet_obj["ATAC"]["label.omic"]

# Load example muscadet object
data(muscadet_obj)
# Access to muscadet omics or slots
muscadet_obj$ATAC
muscadet_obj$genome

# Load example muscadet object
data(muscadet_obj)
# Access to muscomic slots
muscadet_obj$ATAC$label.omic

# Load example muscadet object
data(muscadet_obj)
# Assign new data in muscadet object
muscadet_obj["genome"] <- "hg38"

# Load example muscadet object
data(muscadet_obj)
# Assign new data in muscomic object
muscadet_obj["ATAC"]["label.omic"] <- "scATAC-seq"

# Load example muscadet object
data(muscadet_obj)
# Assign new data in muscadet object
muscadet_obj$genome <- "hg38"

# Load example muscadet object
data(muscadet_obj)
# Assign new data in muscomic object
muscadet_obj$ATAC$label.omic <- "scATAC-seq"

}
