% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet_objects.R
\name{.DollarNames.muscadet}
\alias{.DollarNames.muscadet}
\alias{.DollarNames.muscomic}
\title{Autocompletion for \code{$} access on \code{muscadet} or \code{muscomic} objects}
\usage{
\method{.DollarNames}{muscadet}(x, pattern = "")

\method{.DollarNames}{muscomic}(x, pattern = "")
}
\arguments{
\item{x}{A \code{\link{muscadet}} or \code{\link{muscomic}} object.}

\item{pattern}{ A regular expression.  Only matching names are
    returned.
  }
}
\value{
Character vector of matching element names.
}
\description{
Enable autocompletion for \code{$} access for \code{\link{muscadet}}
or \code{\link{muscomic}} objects. For \code{muscadet} objects, it also lists
omic datasets contained inside the \code{omics} slot.
}
\examples{
# Load muscadet object
data(muscadet_obj)

# Load muscadet object
data(muscadet_obj)

}
