% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet_objects.R
\docType{class}
\name{muscadet-class}
\alias{muscadet-class}
\alias{muscadet}
\title{The muscadet class}
\description{
A \code{muscadet} object encapsulates data from different single-cell omics
as \code{\link{muscomic}} objects as well as downstream analysis results after
clustering and CNA calling.
}
\section{Slots}{

\describe{
\item{\code{omics}}{List of \code{\link{muscomic}} objects, one per single-cell omic (\code{list}).}

\item{\code{bulk.data}}{List of objects containing data from paired bulk sequencing
(\code{list}).}

\item{\code{clustering}}{List of objects containing data associated with the
clustering of cells based on coverage log R ratio values (\code{list}).}

\item{\code{cnacalling}}{List of objects containing data associated with the calling
of copy number alterations (CNAs) (\code{list}).}

\item{\code{genome}}{Reference genome name among: "hg38", "hg19" and "mm10" (\code{character}).}
}}

