% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{muscadet-methods}
\alias{muscadet-methods}
\alias{muscomic-methods}
\alias{coordFeatures}
\alias{matCounts}
\alias{matLogRatio}
\alias{Cells.muscomic}
\alias{Cells.muscadet}
\alias{Features.muscomic}
\alias{Features.muscadet}
\alias{coordFeatures,muscomic-method}
\alias{coordFeatures,muscadet-method}
\alias{matCounts,muscomic-method}
\alias{matCounts,muscadet-method}
\alias{matLogRatio,muscomic-method}
\alias{matLogRatio,muscadet-method}
\title{Methods for \code{\link[=muscomic-class]{muscomic}} and \code{\link[=muscadet-class]{muscadet}}
objects}
\usage{
coordFeatures(x)

matCounts(x)

matLogRatio(x)

\method{Cells}{muscomic}(x, ...)

\method{Cells}{muscadet}(x, ...)

\method{Features}{muscomic}(x, ...)

\method{Features}{muscadet}(x, ...)

\S4method{coordFeatures}{muscomic}(x)

\S4method{coordFeatures}{muscadet}(x)

\S4method{matCounts}{muscomic}(x)

\S4method{matCounts}{muscadet}(x)

\S4method{matLogRatio}{muscomic}(x)

\S4method{matLogRatio}{muscadet}(x)
}
\arguments{
\item{x}{A \code{\link[=muscomic-class]{muscomic}} or \code{\link[=muscadet-class]{muscadet}}
object.}

\item{...}{Other arguments (ignored).}
}
\value{
\code{Cells}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a vector of cell names.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a list of cell names vectors,
one list element per omic.
}

\code{Features}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a vector of feature names.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a list of feature names vectors,
one list element per omic.
}

\code{coordFeatures}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a data frame of feature coordinates.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a list of feature coordinates data frames,
one list element per omic.
}

\code{matCounts}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} \emph{features x cells}.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a list of \code{\link[Matrix:dgCMatrix-class]{dgCMatrices}}
\emph{features x cells}, one list element per omic.
}

\code{matLogRatio}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a \code{matrix} \emph{features x cells}.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a list of \code{matrices} \emph{features x cells},
one list element per omic.
}
}
\description{
Methods to facilitate access to data within the \code{\link[=muscomic-class]{muscomic}}
and \code{\link[=muscadet-class]{muscadet}} objects.
\itemize{
\item \code{Cells()}: Get cell identifiers (addition of methods for \code{muscomic} and
\code{muscadet} to \code{\link[SeuratObject:Cells]{SeuratObject::Cells()}}).
\item \code{Features()}: Get feature identifiers (addition of methods for \code{muscomic}
and \code{muscadet} to \code{\link[SeuratObject:Features]{SeuratObject::Features()}}).
\item \code{coordFeatures()}: Get coordinates of features data frames.
\item \code{matCounts()}: Get raw count matrices.
\item \code{matLogRatio()}: Get log R ratio matrices.
}
}
\details{
The \code{Cells()}, \code{Features()}, and \code{coordFeatures()} functions return information
associated with the coverage slot of \code{muscomic} objects. Their behavior depends
on whether log R ratios have been computed:
\itemize{
\item \strong{Before log R ratio computation:} they return data from the \emph{counts} layer
of the coverage slot.
\item \strong{After log R ratio computation:} if the coverage slot contains a \emph{logratio}
layer, they return the corresponding data instead of the raw counts.
}
}
\examples{
library("SeuratObject")

# Load example muscadet object
# data("exdata_muscadet")

Cells(exdata_muscadet) # list of 2 cell names vectors for the 2 omics
Cells(exdata_muscadet)$ATAC # cell names vector from the omic ATAC
Cells(exdata_muscadet$ATAC) # cell names vector from the ATAC muscomic object
length(Cells(exdata_muscadet))
length(Cells(exdata_muscadet)$ATAC)
length(Cells(exdata_muscadet$ATAC))

Features(exdata_muscadet)
Features(exdata_muscadet)$ATAC
Features(exdata_muscadet$ATAC)

coordFeatures(exdata_muscadet)
coordFeatures(exdata_muscadet)$ATAC
coordFeatures(exdata_muscadet$ATAC)

matCounts(exdata_muscadet)
matCounts(exdata_muscadet)$ATAC
matCounts(exdata_muscadet$ATAC)

matLogRatio(exdata_muscadet)
matLogRatio(exdata_muscadet)$ATAC
matLogRatio(exdata_muscadet$ATAC)

}
\seealso{
\link[=muscomic-class]{muscomic}, \link[=muscadet-class]{muscadet}
}
