% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet-package.R
\docType{package}
\name{muscadet-package}
\alias{muscadet-package}
\title{muscadet: Multiomics Single-Cell Copy Number Alterations Detection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Copy number alterations (CNAs) inference in cancer cells from single-cell multiomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ICAGEN/muscadet}
  \item \url{https://icagen.github.io/muscadet/}
  \item Report bugs at \url{https://github.com/ICAGEN/muscadet/issues}
}

}
\author{
\strong{Maintainer}: Marie Denoulet \email{marie.denoulet@univ-nantes.fr} (\href{https://orcid.org/0000-0002-9668-0439}{ORCID})

Authors:
\itemize{
  \item Marie Denoulet \email{marie.denoulet@univ-nantes.fr} (\href{https://orcid.org/0000-0002-9668-0439}{ORCID})
  \item Nils Giordano \email{nils.giordano@univ-nantes.fr} (\href{https://orcid.org/0000-0003-2549-6631}{ORCID})
  \item Stéphane Minvielle \email{stephane.minvielle@univ-nantes.fr} (\href{https://orcid.org/0000-0003-1389-312X}{ORCID})
  \item Céline Vallot \email{celine.vallot@curie.fr} (\href{https://orcid.org/0000-0003-1601-2359}{ORCID})
  \item Eric Letouzé \email{eric.letouze@gmail.com} (\href{https://orcid.org/0000-0002-6369-2839}{ORCID})
}

Other contributors:
\itemize{
  \item Mia Cherkaoui \email{mia.cherkaoui@univ-nantes.fr} (\href{https://orcid.org/0000-0001-5488-2238}{ORCID}) [contributor]
  \item Florence Magrangeas \email{florence.magrangeas@chu-nantes.fr} (\href{https://orcid.org/0000-0002-5031-3748}{ORCID}) [contributor]
}

}
\keyword{internal}
