% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{MUSCADET_DPP_MODELS}
\alias{MUSCADET_DPP_MODELS}
\title{Available Parametric Determinantal Point Processes}
\usage{
MUSCADET_DPP_MODELS()
}
\value{
A character vector storing the available parametric DPP models.
}
\description{
Lists available DPP parametric models in the package.
}
\examples{
MUSCADET_DPP_MODELS()
}
