% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet_objects.R
\docType{class}
\name{muscomic-class}
\alias{muscomic-class}
\alias{muscomic}
\title{The muscomic class}
\description{
A \code{muscomic} object encapsulates coverage and allelic data from one
single-cell omic dataset used as primary input for muscadet analysis.
}
\section{Slots}{

\describe{
\item{\code{type}}{Type of single-cell omic. Either ATAC or RNA currently supported
(\code{character}).}

\item{\code{label.omic}}{Label to display for the single-cell omic (\code{character}).}

\item{\code{coverage}}{Coverage data on features (\code{list}).}

\item{\code{allelic}}{Allelic data at single nucleotide polymorphisms (SNPs) positions (\code{list}).}
}}

\seealso{
Functions related to \code{muscomic} objects:
\itemize{
\item \code{\link[=CreateMuscomicObject]{CreateMuscomicObject()}} (create \code{muscomic} objects)
\item \link{muscadet-access} (simplified access and assignment methods)
\item \link{.DollarNames.muscadet} (autocompletion for \code{$})
\item \code{\link{show,muscomic-method}} (\code{show} method)
}
}
