% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{my_mle}
\alias{my_mle}
\title{Title}
\usage{
my_mle(
  points,
  lower_bound,
  upper_bound,
  model = "gauss",
  optimizer = "bobyqa",
  marks = NULL,
  num_threads = 1,
  N = 512,
  use_verbose = FALSE
)
}
\arguments{
\item{use_verbose}{}
}
\value{

}
\description{
Title
}
