% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCNA}
\alias{plotCNA}
\title{Plot CNA segments across clusters from a muscadet object}
\usage{
plotCNA(
  x,
  title = NULL,
  cna.colors = c(gain = "#EF6F6AFF", loss = "#6699CCFF", cnloh = "#44AA99FF")
)
}
\arguments{
\item{x}{A \code{\link{muscadet}} object containing CNA calling data to be
visualized (generated using \code{\link[=cnaCalling]{cnaCalling()}}).}

\item{title}{An optional title for the plot. Default is \code{NULL}.}

\item{cna.colors}{A vector of 3 colors for CNA states: gain, loss, and cnloh
(or named vector where names are "gain", "loss", and "cnloh" and the values
are their respective colors). Default is \code{c("gain" = "#EF6F6AFF", "loss" = "#6699CCFF", "cnloh" = "#44AA99FF")}.}
}
\value{
A ggplot object representing the CNA segments plot.
}
\description{
This function visualizes copy number alteration (CNA) segments across
clusters based on data stored in a \code{\link{muscadet}} object. It displays
CNAs for each clusters and scales the y-axis based on the proportion of cells
in each cluster.
}
\examples{
library("ggplot2")

# Load example muscadet object
data(muscadet_obj)

# plot CNA segments
plot <- plotCNA(muscadet_obj, title = "Copy Number Alterations in Example Data")
print(plot)
}
