% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSil}
\alias{plotSil}
\title{Silhouette plot for clustering validation}
\usage{
plotSil(x, k, colors = NULL, title = NULL)
}
\arguments{
\item{x}{A \code{\link{muscadet}} object containing clustering data (using
\code{\link[=clusterMuscadet]{clusterMuscadet()}}).}

\item{k}{Integer specifying the number of clusters to plot (\code{integer}).
It should be within the range of k used for clustering (with
\code{\link[=clusterMuscadet]{clusterMuscadet()}}).}

\item{colors}{Vector of colors for the cluster annotation (\code{character}
vector). Default is \code{NULL}, which uses predefined colors.}

\item{title}{Character string for the title of the plot (\code{character}
string). If \code{NULL}, a default title is generated.}
}
\value{
A ggplot object representing the silhouette plot.
}
\description{
Generate a silhouette plot for a specified clustering result within a
muscadet object.
}
\examples{
# Load a muscadet object
data(muscadet_obj)
plotSil(muscadet_obj, k = 3)

}
