% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP Coordinates from a muscadet Object}
\usage{
plotUMAP(
  x,
  partition = NULL,
  clusters = NULL,
  colors = NULL,
  title = "",
  lab.x = "UMAP 1",
  lab.y = "UMAP 2",
  add_clusters_labels = FALSE,
  point.size = 0.5,
  legend.point.size = 3,
  ...
)
}
\arguments{
\item{x}{A \code{\link[=muscadet-class]{muscadet}} object containing clustering data (using
\code{\link[=clusterMuscadet]{clusterMuscadet()}}).}

\item{partition}{(Optional) Value specifying the clustering partition to
plot (\code{numeric} or \code{character}). It should be either the resolution or the
number of cluster (k) used for clustering depending on the clustering
method (\code{res_range} or \code{k_range} with \code{\link[=clusterMuscadet]{clusterMuscadet()}}).
If both \code{partition} and \code{clusters} arguments are \code{NULL} (default), the
assigned clusters for CNA calling (\code{x@cnacalling$cluster}) are used if
available in \code{x} (see \code{\link[=assignClusters]{assignClusters()}}).}

\item{clusters}{(Optional) A custom named vector of cluster assignments
(\code{integer} named vector). Names must corresponds to cell names within the
muscadet object \code{x}. If it contains less cells than the muscadet object
\code{x}, the missing cells are filtered out and not displayed in the heatmap.
If \code{show_missing = FALSE} only the provided cells with data in all omics
will be displayed. If both \code{partition} and \code{clusters} arguments are \code{NULL}
(default), the assigned clusters for CNA calling (\code{x@cnacalling$cluster})
are used if available in \code{x} (see \code{\link[=assignClusters]{assignClusters()}}).}

\item{colors}{Vector of colors for the cluster annotation (\code{character}
vector). If \code{NULL} (default), it uses predefined colors.}

\item{title}{Character string for the title of the plot (\code{character}
string). Default is an empty character string.}

\item{lab.x}{Label for the x-axis (\code{character} string). Default is "UMAP 1".}

\item{lab.y}{Label for the y-axis (\code{character} string). Default is "UMAP 2".}

\item{add_clusters_labels}{Logical. If \code{TRUE}, adds the cluster names as
text or boxed labels using \code{\link[=add_labels]{add_labels()}}. Default is \code{FALSE}.}

\item{point.size}{Numeric. Size of the points in the UMAP plot passed to
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}. Default is \code{0.5}.}

\item{legend.point.size}{Numeric. Size of the points in the legend of the
UMAP plot. Default is \code{3}.}

\item{...}{Additional arguments passed to \code{\link[=add_labels]{add_labels()}} providing an
underlying geom for label names (\code{\link[ggplot2:geom_text]{geom_text}},
\code{\link[ggplot2:geom_label]{geom_label}},
\code{\link[ggrepel:geom_text_repel]{geom_text_repel}}, or
\code{\link[ggrepel:geom_label_repel]{geom_label_repel}}).}
}
\value{
A \code{ggplot} object.
}
\description{
Visualize UMAP (Uniform Manifold Approximation and Projection) coordinates
stored in a \code{\link[=muscadet-class]{muscadet}} object. The function allows coloring
by clusters and optionally adding cluster labels.
}
\examples{
\dontrun{

# Load example muscadet object
# data("exdata_muscadet")

p <- plotUMAP(exdata_muscadet, partition = 0.3, title = "UMAP copy-number clusters")
print(p)
}

}
