% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaCalling.R
\name{preProcSample2}
\alias{preProcSample2}
\title{Process read count matrix and segmentation}
\source{
This function is derived from the \code{\link[facets:preProcSample]{facets::preProcSample()}} function,
with the modifications to fit its use in muscadet.

Seshan VE, Shen R (2021). \emph{facets: Cellular Fraction and Copy Numbers from Tumor Sequencing}.
R package version 0.6.2, \url{https://github.com/mskcc/facets}.
}
\usage{
preProcSample2(
  rcmat,
  het.thresh = 0.25,
  snp.nbhd = 250,
  cval = 25,
  gbuild = "hg38",
  hetscale = TRUE,
  ndepth = 5,
  ndepthmax = 1000
)
}
\arguments{
\item{rcmat}{A data frame with 8 required columns: \code{Chrom}, \code{Pos}, \code{NOR.DP},
\code{NOR.RD}, \code{TUM.DP}, \code{TUM.RD}, \code{cluster}, and \code{signal} (\code{data.frame}).}

\item{het.thresh}{VAF (Variant Allele Frequency) threshold to call variant positions
heterozygous (\code{numeric}). Default: 0.25.}

\item{snp.nbhd}{Window size for selecting loci to reduce serial correlation
(\code{numeric}). Default: 250.}

\item{cval}{Critical value for segmentation (\code{numeric}). Default: 25.}

\item{gbuild}{Genome build used for alignment. One of \code{"hg19"},
\code{"hg38"}, or \code{"mm10"} (\code{character}). Default: \code{"hg19"}.}

\item{hetscale}{Logical value indicating whether log odds ratio (logOR)
should be scaled to give more weight in the test statistics for
segmentation and clustering (\code{logical}). Usually only 10\% of snps are hets
and hetscale gives the logOR contribution to T-square as 0.25/proportion of
hets. Default: \code{TRUE}.}

\item{ndepth}{Minimum depth in normal reference to keep (\code{numeric}). Default:
5.}

\item{ndepthmax}{Maximum normal coverage threshold for filtering loci
(\code{numeric}). Default: 1000.}
}
\value{
A list containing:
\describe{
\item{pmat}{Read counts and other elements of all the loci.}
\item{gbuild}{Genome build used for the analysis.}
\item{nX}{Chromosome number for X (e.g., 23 for human, 20 for mouse).}
\item{clusters}{Unique clusters from the processed data.}
\item{seg.tree}{Segmentation tree for each chromosome.}
\item{jointseg}{Segmented variant positions data.}
\item{hscl}{Scaling factor for logOR data.}
}
}
\description{
Function adapted from \code{\link[facets:preProcSample]{facets::preProcSample()}} to process read count matrix
and generates a segmentation tree. The modifications from the original
function includes:
\itemize{
\item Incorporation of \code{cluster} and \code{signal} columns into the final result.
\item Change in the log ratio (\code{cnlr}) for allelic data, it is computed as the
log ratio (\code{cnlr}) mean between the previous and next coverage positions.
}
}
\details{
The function processes variant positions data to generate a segmentation
tree. It uses \code{\link[facets]{procSnps}} to compute initial values,
adjusts the log ratio for allelic signals, and computes segmentation using
\code{\link[facets]{segsnps}}.

SNPs (or variants) in a genome are not evenly spaced, and loci are sampled
within the specified window (\code{snp.nbhd}) to reduce serial correlation.
}
\examples{
library("facets")

# Load example muscadet object
data(muscadet_obj)

counts <- muscadet_obj$cnacalling$combined.counts
counts <- counts[complete.cases(counts),]
counts_clus <- counts[which(counts$cluster == 1),]
result <- preProcSample2(counts_clus)

}
\references{
\describe{
\item{\link{facets-package} package}{Shen R, Seshan VE. FACETS: allele-specific copy number and
clonal heterogeneity analysis tool for high-throughput DNA sequencing.
Nucleic Acids Res. 2016 Sep 19;44(16):e131.
doi: \href{https://www.doi.org/10.1093/nar/gkw520}{10.1093/nar/gkw520}.
PMID: 27270079; PMCID: PMC5027494.}
}
}
\seealso{
\code{\link[facets]{preProcSample}}, \code{\link[facets]{procSnps}},
\code{\link[facets]{segsnps}}
}
