% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_allele}
\alias{process_allele}
\title{Process allele counts results from SCReadCounts format}
\usage{
process_allele(data, vcf = NULL, samplename = NULL)
}
\arguments{
\item{data}{A data frame of SCReadCounts output table (\code{data.frame}).}

\item{vcf}{An optional VCF-format data frame with unphased/phased genotype
information to add it in output (\code{data.frame}).}

\item{samplename}{An optional sample name to prefix cell barcodes
(\code{character}).}
}
\value{
A cleaned data frame with VCF-like standardized columns:
\describe{
\item{\code{cell}}{Barcodes of cells (\code{character}).}
\item{\code{id}}{Variant unique identifier defined as
CHROM_POS_REF_ALT, e.g. "1_920949_C_G" (\code{character}).}
\item{\code{CHROM}}{Chromosome in integer format, e.g. 15 (X and Y chromosomes
are not included) (\code{integer}).}
\item{\code{POS}}{Position of the variant (1-base positions) (\code{integer}).}
\item{\code{REF}}{Reference allele base, "A" "C" "G" or "T" (\code{character}).}
\item{\code{ALT}}{Alternative allele base, "A" "C" "G" or "T" (\code{character}).}
\item{\code{RD}}{Reference allele depth/count (\code{integer}).}
\item{\code{AD}}{Alternative allele depth/count (\code{integer}).}
\item{\code{DP}}{Total depth/count (\code{integer}).}
\item{\code{GT}}{Genotype (if \code{vcf} provided): "0/1" or "1/0" if unphased; "0|1"
or "1|0" if phased. (\code{character}).}
}
}
\description{
This function reformats SCReadCounts output into a standardized VCF-like data
frame and optionally merges unphased/phased genotype information from a
VCF-format data frame.
}
\note{
\itemize{
\item \href{https://horvathlab.github.io/NGS/SCReadCounts/}{SCReadCounts website}
\item \href{https://en.wikipedia.org/wiki/Variant_Call_Format}{Variant Call Format (VCF) format}
}
}
\examples{

# Example data
sc_data <- data.frame(
  ReadGroup = c("cell1", "cell2"),
  CHROM = c(1, 1),
  POS = c(1001, 1002),
  REF = c("A", "C"),
  ALT = c("G", "T"),
  SNVCount = c(3, 5),
  RefCount = c(7, 5),
  GoodReads = c(10, 10)
)

vcf_data <- data.frame(
  CHROM = c(1, 1),
  POS = c(1001, 1002),
  ID = c(".", "."),
  REF = c("A", "C"),
  ALT = c("G", "T"),
  QUAL = c(".", "."),
  FILTER = c(".", "."),
  INFO = c(".", "."),
  FORMAT = c("GT", "GT"),
  sample1 = c("0|1", "1|0"),
  stringsAsFactors = FALSE
)

data <- process_allele(sc_data, vcf = vcf_data, samplename = "sampleA")

}
\references{
\href{https://horvathlab.github.io/NGS/SCReadCounts/}{SCReadCounts}

Liu, H., Bousounis, P., Movassagh, M., Edwards, N., and Horvath, A.
SCReadCounts: estimation of cell-level SNVs expression from scRNA-seq data.
BMC Genomics 22, 689 (2021).
doi: \href{https://doi.org/10.1186/s12864-021-07974-8}{10.1186/s12864-021-07974-8}
}
