% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{rbidpp}
\alias{rbidpp}
\title{Random generation of point patterns from a cross-type DPP}
\usage{
rbidpp(
  n = 1,
  rho1 = 100,
  rho2 = 100,
  tau = 0.2,
  alpha1 = 0.03,
  alpha2 = 0.03,
  alpha12 = 0.05,
  L = 1,
  seed = 1234,
  progress = TRUE,
  ncores = parallelly::availableCores(omit = 1)
)
}
\arguments{
\item{n}{An integer value specifying the sample size. Defaults to \code{1L}.}

\item{rho1}{A numeric value specifying the intensity of the 1st marginal DPP.
Defaults to \code{100L}.}

\item{rho2}{A numeric value specifying the intensity of the 2nd marginal DPP.
Defaults to \code{100L}.}

\item{tau}{A numeric value specifying the correlation between marks. Defaults
to \code{0.2}.}

\item{alpha1}{A numeric value specifying the repulsion rate of the 1st DPP.
Defaults to \code{0.03}.}

\item{alpha2}{A numeric value specifying the repulsion rate of the 2nd DPP.
Defaults to \code{0.03}.}

\item{alpha12}{A numeric value specifying the repulsion rate between marks.
Defaults to \code{0.05}.}

\item{L}{An integer value specifying the window size of the point pattern.
Defaults to \code{1L}.}

\item{seed}{An integer value specifying the random generator seed. Defaults
to \code{1234}.}

\item{progress}{A boolean value specifying whether a progress bar should be
displayed. Defaults to \code{TRUE}.}

\item{ncores}{An integer value specifying the number of cores to use.
Defaults to all cores available except one.}
}
\value{
An object of class \link[spatstat.geom:ppp]{spatstat.geom::ppp} storing the simulated point
pattern.
}
\description{
Simulates one or more planar point patterns following a given multi-mark DPP.
}
\examples{
#pp <- rbidpp()
}
