% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_as_vcf}
\alias{save_as_vcf}
\title{Save a data frame as a VCF File}
\usage{
save_as_vcf(data, file, header = NULL)
}
\arguments{
\item{data}{A data frame containing variant data (\code{data.frame}). The first
column should represent the chromosome.}

\item{file}{A character string specifying the output file path with the
\code{.vcf} or \code{.vcf.gz} extension (\verb{character string}). If the file has the
\code{.vcf.gz} extension it will be compressed.}

\item{header}{An optional custom header for the VCF format (\code{character}
vector). By default, \code{NULL} sets a minimal header.}
}
\value{
The function does not return a value but writes a VCF file to the specified path.
}
\description{
This function saves a given data frame in \href{https://en.wikipedia.org/wiki/Variant_Call_Format}{variant calling format (VCF)} with necessary
headers.
}
\details{
The function ensures that the first column is named \verb{#CHROM} and writes standard VCF headers.
It then appends the variant data in a tab-separated format.
}
\examples{
\dontrun{
vcf_data <- data.frame(
  CHROM = c("chr1", "chr2"),
  POS = c(12345, 67890),
  ID = c(".", "."),
  REF = c("A", "T"),
  ALT = c("G", "C"),
  QUAL = c(".", "."),
  FILTER = c("PASS", "PASS"),
  INFO = c("AF=0.5", "AF=0.3"),
  FORMAT = c("GT", "GT"),
  SAMPLE = c("0/1", "1/1")
)
save_as_vcf(vcf_data, "output.vcf")
}

}
