% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{show,muscadet-method}
\alias{show,muscadet-method}
\title{muscadet object overview}
\usage{
\S4method{show}{muscadet}(object)
}
\arguments{
\item{object}{A \code{\link[=muscadet-class]{muscadet}} object.}
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly returns
\code{NULL}
}
\description{
Overview of a \code{\link[=muscadet-class]{muscadet}} object.
}
\examples{
# Load example muscadet object
# data("exdata_muscadet")

# Overview of the muscadet object
show(exdata_muscadet)

# Overview of the muscomic objects within
show(slot(exdata_muscadet, "omics"))

# Overview of the first muscomic objects within
show(slot(exdata_muscadet, "omics")[[1]])

}
\seealso{
\link[=muscadet-class]{muscadet}, \code{\link[=CreateMuscadetObject]{CreateMuscadetObject()}}
}
\keyword{internal}
