% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{show,muscomic-method}
\alias{show,muscomic-method}
\title{muscomic object overview}
\usage{
\S4method{show}{muscomic}(object)
}
\arguments{
\item{object}{A \code{\link[=muscomic-class]{muscomic}} object.}
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly returns
\code{NULL}
}
\description{
Overview of a \code{\link[=muscomic-class]{muscomic}} object.
}
\examples{
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = exdata_mat_counts_atac_tumor,
  allele_counts = exdata_allele_counts_atac_tumor,
  features = exdata_peaks
)
atac
}
\seealso{
\link[=muscomic-class]{muscomic}, \code{\link[=CreateMuscomicObject]{CreateMuscomicObject()}}
}
\keyword{internal}
