% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_bessel4}
\alias{sim_bessel4}
\title{Simulated Data from Two-Mark Bessel DPP With 0.4 Correlation}
\format{
\subsection{\code{sim_bessel4}}{

A \link[base:list]{base::list} with 100 entries, each of them storing a planar point pattern
following a two-mark Bessel DPP on a window size \eqn{L=1} with \eqn{\rho_1 =
100}, \eqn{\rho_2 = 100}, \eqn{\alpha_1 = 0.03}, \eqn{\alpha_2 = 0.03},
\eqn{\alpha_{12} = 0.035} and \eqn{\tau = 0.4}.
}
}
\source{
The data set has been generated via \link{rbidpp}.
}
\usage{
sim_bessel4
}
\description{
A list of 100 simulated point patterns according to the two-mark Bessel DPP
with a correlation of \eqn{0.4}.
}
\keyword{datasets}
