% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_gauss3}
\alias{sim_gauss3}
\title{Simulated Data from Two-Mark Gaussian DPP With 0.3 Correlation}
\format{
\subsection{\code{sim_gauss3}}{

A \link[base:list]{base::list} with 100 entries, each of them storing a planar point pattern
following a two-mark Gaussian DPP on a window size \eqn{L=1} with \eqn{\rho_1
= 100}, \eqn{\rho_2 = 100}, \eqn{\alpha_1 = 0.03}, \eqn{\alpha_2 = 0.03},
\eqn{\alpha_{12} = 0.035} and \eqn{\tau = 0.3}.
}
}
\source{
The data set has been generated via \link{rbidpp}.
}
\usage{
sim_gauss3
}
\description{
A list of 100 simulated point patterns according to the two-mark Gaussian DPP
with a correlation of \eqn{0.3}.
}
\keyword{datasets}
