% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern_functions.R
\name{simdppxmatern}
\alias{simdppxmatern}
\title{Simulate points according to a bivariate DPP}
\usage{
simdppxmatern(rho1 = 100, rho2 = 100, tau = 0.2, alpha1 = 0.03,
  alpha2 = 0.03, alpha12 = 0.05, nu1 = 10, nu2 = 10, nu12 = 10,
  progress = 0, Kspec = "Kspecmatern", testtau = "testtaumatern")
}
\arguments{
\item{rho1}{A numeric scalar specifying the intensity of the 1st DPP (default: 100).}

\item{rho2}{A numeric scalar specifying the intensity of the 2nd DPP (default: 100).}

\item{tau}{A numeric scalar specifying the cross-correlation (default: 0.2).}

\item{alpha1}{A numeric scalar specifying the ??? of the 1nd DPP (default: 0.03).}

\item{alpha2}{A numeric scalar specifying the ??? of the 2nd DPP (default: 0.03).}

\item{alpha12}{A numeric scalar specifying the cross-??? (default: 0.05).}

\item{nu1}{A numeric scalar specifying the ??? of the first DPP (default: 10).}

\item{nu2}{A numeric scalar specifying the ??? of the 2nd DPP (default: 10).}

\item{nu12}{A numeric scalar specifying the cross-??? (default: 10).}

\item{progress}{A numeric specifying whether a progess bar should be displayed (default: 0 for none).}

\item{Kspec}{A function specifying the kernel to be used (default: \code{Kspecbessel}).}

\item{testtau}{A function specifying the upper bound for the cross-correlation (default: \code{testtaubessel}).}
}
\value{
A \code{\link[spatstat]{ppp}} object containing the simulated point pattern.
}
\description{
Simulate points according to a bivariate DPP
}
\examples{
pp <- simdppxmatern(progress = 1, Kspec = "Kspecbessel", testtau = "testtaubessel")
}
