% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulations}
\alias{simulations}
\title{Simulated Data from Two-Mark DPPs}
\format{
\subsection{\code{simulations}}{

A \link[tibble:tibble]{tibble::tibble} with 9 rows and 9 columns:
\describe{
\item{window_size}{Window size of the point pattern;}
\item{rho1, rho2, alpha1, alpha2, alpha12, tau, k12}{Parameter set of the
two-mark Gaussian DPP;}
\item{sim}{List of 100 simulated data sets according to the multi-mark
Gaussian DPP with correspoonding parameter set.}
}
}
}
\source{
The data set has been generated via \link{rbidpp}.
}
\usage{
simulations
}
\description{
A collection of simulated simulated point patterns according to the two-mark
Gaussian DPP with various sets of parameters.
}
\keyword{datasets}
