% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-mark-dpp.R
\name{TwoMarkDPP}
\alias{TwoMarkDPP}
\title{Two-Mark Determinantal Point Process Model}
\description{
Implements a generic two-mark planar determinantal point process
model. It is possible to set its parameters, fit the model to an observed
planar point pattern or simulate a planar point pattern.
}
\examples{

## ------------------------------------------------
## Method `TwoMarkDPP$new`
## ------------------------------------------------

mod <- TwoMarkDPP$new(model = "Gauss")

## ------------------------------------------------
## Method `TwoMarkDPP$fit`
## ------------------------------------------------

mod <- TwoMarkDPP$new(model = "Gauss")
mod$fit(sim_gauss0[[1]])

## ------------------------------------------------
## Method `TwoMarkDPP$random`
## ------------------------------------------------

mod <- TwoMarkDPP$new(model = "Gauss")
mod$fit(sim_gauss0[[1]])
mod$random(n = 1)

## ------------------------------------------------
## Method `TwoMarkDPP$print`
## ------------------------------------------------

mod <- TwoMarkDPP$new()
mod
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{first_intensity}}{Setter and getter for the first intensity
\eqn{\rho_1}.}

\item{\code{second_intensity}}{Setter and getter for the second intensity
\eqn{\rho_2}.}

\item{\code{first_repulsion_rate}}{Setter and getter for the first repulsion
rate \eqn{\alpha_1}.}

\item{\code{second_repulsion_rate}}{Setter and getter for the second repulsion
rate \eqn{\alpha_2}.}

\item{\code{cross_repulsion_rate}}{Setter and getter for the cross repulsion
rate \eqn{\alpha_{12}}.}

\item{\code{between_mark_correlation}}{Setter and getter for the correlation
\eqn{\tau} between marks.}

\item{\code{window_size}}{Setter and getter for the window size \eqn{L}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TwoMarkDPP-new}{\code{TwoMarkDPP$new()}}
\item \href{#method-TwoMarkDPP-fit}{\code{TwoMarkDPP$fit()}}
\item \href{#method-TwoMarkDPP-random}{\code{TwoMarkDPP$random()}}
\item \href{#method-TwoMarkDPP-print}{\code{TwoMarkDPP$print()}}
\item \href{#method-TwoMarkDPP-clone}{\code{TwoMarkDPP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoMarkDPP-new"></a>}}
\if{latex}{\out{\hypertarget{method-TwoMarkDPP-new}{}}}
\subsection{Method \code{new()}}{
The \link{TwoMarkDPP} constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoMarkDPP$new(model = AVAILABLE_MODELS)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A string specifying the underlying parametric model to be
used. Choices are \code{"Gauss"} or \code{"Bessel"}. Defaults to \code{"Gauss"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \link{TwoMarkDPP} object invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mod <- TwoMarkDPP$new(model = "Gauss")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoMarkDPP-fit"></a>}}
\if{latex}{\out{\hypertarget{method-TwoMarkDPP-fit}{}}}
\subsection{Method \code{fit()}}{
Estimates the parameters of the two-mark DPP from an
observed planar point pattern.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoMarkDPP$fit(X, method = c("pcf", "mle"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{An object of class \link[spatstat.geom:ppp]{spatstat.geom::ppp}.}

\item{\code{method}}{A string specifying the kind of estimator to use. Choices
are either \code{"pcf"} or \code{"mle"}. Defaults to \code{"pcf"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The \link{TwoMarkDPP} object invisibly.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mod <- TwoMarkDPP$new(model = "Gauss")
mod$fit(sim_gauss0[[1]])
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoMarkDPP-random"></a>}}
\if{latex}{\out{\hypertarget{method-TwoMarkDPP-random}{}}}
\subsection{Method \code{random()}}{
Samples one or more planar point patterns according to the
two-mark DPP.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoMarkDPP$random(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{An integer value specifying the sample size.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[base:list]{base::list} of objects of class \link[spatstat.geom:ppp]{spatstat.geom::ppp} storing
sampled planar point patterns.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mod <- TwoMarkDPP$new(model = "Gauss")
mod$fit(sim_gauss0[[1]])
mod$random(n = 1)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoMarkDPP-print"></a>}}
\if{latex}{\out{\hypertarget{method-TwoMarkDPP-print}{}}}
\subsection{Method \code{print()}}{
Prints the \link{TwoMarkDPP} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoMarkDPP$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Extra parameters to be passed on to next methods. Current
unused.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{mod <- TwoMarkDPP$new()
mod
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TwoMarkDPP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TwoMarkDPP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TwoMarkDPP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
