% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allele_counts}
\alias{allele_counts}
\alias{allele_counts_atac_tumor}
\alias{allele_counts_atac_ref}
\alias{allele_counts_rna_tumor}
\alias{allele_counts_rna_ref}
\title{Example data: Allele counts at variation positions}
\format{
A data frame with columns based on the \href{https://en.wikipedia.org/wiki/Variant_Call_Format}{Variant Call Format (VCF) format} columns.
It contains the following columns:
\describe{
\item{\code{cell}}{Barcodes of cells (\code{character}).}
\item{\code{id}}{Variant unique identifier defined as
CHROM_POS_REF_ALT, e.g. "1_920949_C_G" (\code{character}).}
\item{\code{CHROM}}{Chromosome in integer format, e.g. 15 (X and Y chromosomes
are not included) (\code{integer}).}
\item{\code{POS}}{Position of the variant (1-base positions) (\code{integer}).}
\item{\code{REF}}{Reference allele base, "A" "C" "G" or "T" (\code{character}).}
\item{\code{ALT}}{Alternative allele base, "A" "C" "G" or "T" (\code{character}).}
\item{\code{RD}}{Reference allele depth/count (\code{integer}).}
\item{\code{AD}}{Alternative allele depth/count (\code{integer}).}
\item{\code{DP}}{Total depth/count (\code{integer}).}
\item{\code{GT}}{Genotype: "0/1" or "1/0" if unphased; "0|1" or "1|0" if phased.) (\code{character}).}
}
}
\usage{
allele_counts_atac_tumor

allele_counts_atac_ref

allele_counts_rna_tumor

allele_counts_rna_ref
}
\description{
Data frames of allele counts at variant positions per cell. Variant positions
can be either common single nucleotide polymorphisms (SNPs) positions or
individual-specific heterozygous positions retrieved by bulk sequencing.
}
\keyword{datasets}
