% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_seurat}
\alias{cluster_seurat}
\title{Multi Omics Clustering using Seurat Multi Modal Graph-based Clustering}
\usage{
cluster_seurat(
  mat_list,
  res_range = seq(0.1, 0.5, 0.1),
  dims_list = rep(list(1:8), length(mat_list)),
  algorithm = 1,
  knn_seurat = 20,
  knn_range_seurat = 200,
  quiet = FALSE
)
}
\arguments{
\item{mat_list}{A named list of log R ratio matrices (cells x features), one
per omic layer (\code{list}).}

\item{res_range}{A numeric non-negative vector specifying the resolution
values to use for \code{\link[Seurat:FindClusters]{Seurat::FindClusters()}} (\code{numeric} vector). Default is
\code{c(0.1, 0.2, 0.3, 0.4, 0.5)}.}

\item{dims_list}{A list of vectors of PC dimensions to use for each omic
(\code{list}). Must match the length of \code{mat_list} (e.g., list(1:8) for 1 omic ;
list(1:8, 1:8) for 2 omics). Default is the first 8 dimensions for each
provided omic.}

\item{algorithm}{Integer specifying the algorithm for modularity optimization
by \link[Seurat:FindClusters]{Seurat::FindClusters} (\code{1} = original Louvain algorithm; \code{2} = Louvain algorithm with multilevel
refinement; \code{3} = SLM algorithm; \code{4} = Leiden algorithm). Leiden requires the
leidenalg python. Default is \code{1}.}

\item{knn_seurat}{Integer specifying the number of nearest neighbors used for
graph construction with Seurat functions \code{\link[Seurat:FindNeighbors]{Seurat::FindNeighbors()}}
(\code{k.param}) or \code{\link[Seurat:FindMultiModalNeighbors]{Seurat::FindMultiModalNeighbors()}} (\code{k.nn}) (\code{integer}).
Default is \code{20}.}

\item{knn_range_seurat}{Integer specifying the approximate number of nearest
neighbors to compute for \code{\link[Seurat:FindMultiModalNeighbors]{Seurat::FindMultiModalNeighbors()}} (\code{knn.range})
(\code{integer}). Default is \code{200}.}

\item{quiet}{\code{TRUE} or \code{FALSE} (\code{logical}). Whether to turn off messages. By
default: \code{FALSE}.}
}
\value{
A list containing:
\itemize{
\item \code{params}: List of parameters used for clustering (\code{list}).
\item \code{nn}: Nearest neighbors object (\code{Neighbor} \link[SeuratObject:Neighbor-class]{SeuratObject::Neighbor}).
\item \code{graph}: Shared nearest neighbors graph (\code{Graph} \link[SeuratObject:Graph-class]{SeuratObject::Graph}).
}

\describe{
\item{params}{List of parameters used for clustering (\code{list}).}
\item{nn}{Nearest neighbors object (\code{Neighbor} \link[SeuratObject:Neighbor-class]{SeuratObject::Neighbor}).}
\item{graph}{Shared nearest neighbors graph (\code{Graph} \link[SeuratObject:Graph-class]{SeuratObject::Graph}).}
\item{dist}{Distance matrix derived from the graph (\code{matrix}).}
\item{umap}{UMAP coordinates (\code{matrix}).}
\item{clusters}{A named list of clustering results (vectors of cluster
labels) for each value in \code{res_range} (\code{list}).}
}
}
\description{
Performs graph-based clustering of cells using Seurat, based on one or two
log R ratio matrices (\code{mat_list}), including shared nearest neighbors (SNN)
graph construction on selected dimensions from PCA (\code{dims_list}), to identify
clusters of cells for each specified resolution (\code{res_range}).
}
\details{
\itemize{
\item For two omics: multimodal integration is performed using
\code{\link[Seurat:FindMultiModalNeighbors]{Seurat::FindMultiModalNeighbors()}} (weighted shared nearest neighbors graph).
Only common cells between omics are used.
\item For a single omic: \code{\link[Seurat:FindNeighbors]{Seurat::FindNeighbors()}} (shared nearest neighbors
graph) is used.
}
}
\examples{
data("muscadet_obj", package = "muscadet")
mat_list <- lapply(muscadet::matLogRatio(muscadet_obj), t)
result <- cluster_seurat(mat_list, res_range = seq(0.2, 0.4, 0.1))

#' @examples
# Load example muscadet object
data(muscadet_obj)

# Format input
# transpose matrices to: cells x features matrices
mat_list <- lapply(muscadet::matLogRatio(muscadet_obj), t)

# Run integration & clustering
result <- cluster_seurat(mat_list, res_range = c(0.1, 0.3, 0.5))

# View results
lapply(result$clusters, table)

}
\seealso{
\href{https://satijalab.org/seurat/articles/weighted_nearest_neighbor_analysis}{Weighted Nearest Neighbor Analysis Vignette from Seurat}
}
