% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnaCalling.R
\name{cnaCalling}
\alias{cnaCalling}
\title{Copy Number Alteration (CNA) Calling from muscadet object}
\source{
This function uses several functions from the \link{facets} package,
including: \code{\link[facets:facets-internal]{facets::clustersegs()}}, \code{\link[facets:emcncf]{facets::emcncf()}},
\code{\link[facets:facets-internal]{facets::findDiploidLogR()}}, \code{\link[facets:fitcncf]{facets::fitcncf()}}, \code{\link[facets:procSample]{facets::procSample()}},
\code{\link[facets:facets-internal]{facets::procSnps()}}, and adapted function \code{\link[=preProcSample2]{preProcSample2()}}.

Seshan VE, Shen R (2021). \emph{facets: Cellular Fraction and Copy Numbers from
Tumor Sequencing}. R package version 0.6.2,
\url{https://github.com/mskcc/facets}.
}
\usage{
cnaCalling(
  x,
  omics.coverage = NULL,
  depthmin.a.clusters = 30,
  depthmin.c.clusters = 50,
  depthmin.a.allcells = 30,
  depthmin.c.allcells = 50,
  depthmin.nor = 5,
  depthmax.nor = 1000,
  het.thresh = 0.25,
  snp.nbhd = 250,
  hetscale = TRUE,
  cval1 = 25,
  cval2 = 150,
  min.nhet = 5,
  clonal.thresh = 0.9,
  dist.breakpoints = 1e+06,
  ploidy = "auto",
  quiet = FALSE
)
}
\arguments{
\item{x}{A \code{\link{muscadet}} object. Must contain:
\itemize{
\item Clustering assignments in the \code{cnacalling$clusters} slot
(use \code{\link[=assignClusters]{assignClusters()}}).
\item Combined allelic and coverage counts per cluster in the
\code{cnacalling$combined.counts} slot (use \code{\link[=mergeCounts]{mergeCounts()}}).
}}

\item{omics.coverage}{A vector of omics names to select for coverage log R
ratio data. RECOMMENDED: select "ATAC" when ATAC and RNA omics are
available, the ATAC coverage (DNA) signal is less noisy than RNA signal. By
default, \code{NULL} selects all available data.}

\item{depthmin.a.clusters}{Minimum allelic depth per clusters in tumor cells
(default: 30).}

\item{depthmin.c.clusters}{Minimum coverage depth per clusters in tumor cells
(default: 50).}

\item{depthmin.a.allcells}{Minimum allelic depth for all tumor cells
(default: 30).}

\item{depthmin.c.allcells}{Minimum coverage depth for all tumor cells
(default: 50).}

\item{depthmin.nor}{Minimum coverage depth for normal sample (default: 0).}

\item{depthmax.nor}{Maximum coverage depth for normal sample (default: 1000).}

\item{het.thresh}{VAF (Variant Allele Frequency) threshold to call variant
positions heterozygous for \code{\link[=preProcSample2]{preProcSample2()}} (default: 0.25).}

\item{snp.nbhd}{Window size for selecting SNP loci to reduce serial
correlation for \code{\link[=preProcSample2]{preProcSample2()}} (default: 250).}

\item{hetscale}{Logical value indicating whether log odds ratio (logOR)
should be scaled to give more weight in the test statistics for
segmentation and clustering \code{\link[=preProcSample2]{preProcSample2()}}. (default: \code{TRUE})}

\item{cval1}{Critical value for segmentation for \code{\link[=preProcSample2]{preProcSample2()}} (default:
25).}

\item{cval2}{Critical value for segmentation for \code{\link[facets:procSample]{facets::procSample()}}
(default: 150).}

\item{min.nhet}{Minimum number of heterozygous positions in a segment for
\code{\link[facets:procSample]{facets::procSample()}} and \code{\link[facets:emcncf]{facets::emcncf()}} (default: 5).}

\item{clonal.thresh}{Threshold of minimum cell proportion to label a
segment as clonal (default: 0.9).}

\item{dist.breakpoints}{Minimum distance between breakpoints to define
distinct segments (default: 1e6).}

\item{ploidy}{Specifies ploidy assumption: \code{"auto"}, \code{"median"}, or numeric
value (default: \code{"auto"}).}

\item{quiet}{Logical. If \code{TRUE}, suppresses informative messages during
execution. Default is \code{FALSE}.}
}
\value{
A modified \code{\link{muscadet}} object with added CNA analysis results in
the \code{cnacalling} slot, including: filtered counts and positions, segmentation
data for clusters and all cells, consensus segments across clusters based on
breakpoints, diploid log R ratio, purity and ploidy.

Details of the cnacalling slot:
\itemize{
\item \code{combined.counts.filtered}: Filtered counts per clusters.
\item \code{combined.counts.allcells}: Counts summed for all cells (no
cluster distinction).
\item \code{combined.counts.allcells.filtered}: Filtered counts summed for
all cells (no cluster distinction).

\item \code{positions}: Data frame of positions from the per cluster
analysis. Positions in rows and associated data in columns: \code{chrom},
\code{maploc} (position), \code{rCountT} (read count in tumor), \code{rCountN} (read count
in normal), \code{vafT} (variant allele frequency in tumor), \code{vafN} (variant
allele frequency in normal), \code{cluster} (cluster id), \code{signal} (whether the
counts come from coverage or allelic data), \code{het} (heterozygous status),
\code{keep} (whether to keep position), \code{gcpct} (GC percentage), \code{gcbias} (GC
bias correction), \code{cnlr} (log R ratio), \code{valor} (log odds ratio), \code{lorvar}
(variance of log odds ratio), \code{seg0}, \code{seg_ori} (segment original id within
each cluster), \code{seg} (segment id), \code{segclust} (cluster of segments id),
\code{vafT.allcells} (vairiant allele frequency in all tumor cells), \code{colVAR}
(integer for allelic position color depending on \code{vafT.allcells}).

\item \code{segments}: Data frame of segments from the per cluster
analysis. Segments in rows and associated data in columns: \code{chrom}, \code{seg}
(segment id), \code{num.mark} (number of positions in segment), \code{nhet} (number
of heterezygous positions in segment), \code{cnlr.median} (segment log R ratio
median), \code{mafR} (segment square of expected log odds ratio), vafT.median
(segment variant allele frequency median), \code{cluster} (cluster id),
\code{seg_ori} (segment original id within each cluster), \code{segclust} (cluster of
segments id), \code{cnlr.median.clust} (segment cluster log R ratio median),
\code{mafR.clust} (segment cluster square of expected log odds ratio), \code{cf}
(cell fraction), \code{tcn} (total copy number), \code{lcn} (lower copy number),
\code{start}, \code{end}, \code{cf.em} (cell fraction computed with EM algorithm),
\code{tcn.em}, (total copy number computed with EM algorithm), \code{lcn.em} (lower
copy number computed with EM algorithm).

\item \code{positions.allcells}: Same as \code{positions} but from the all cells analysis.
\item \code{segments.allcells}: Same as \code{segments} but from the all cells analysis.

\item \code{consensus.segs}: Data frame of unique consensus segments across
clusters, with the \code{cna} (\code{logical}) and \code{cna_clonal} (\code{logical})
information.

\item \code{table}: Data frame of consensus segments across clusters with
associated information per cluster in columns: \code{chrom}, \code{start}, \code{end}, \code{id},
\code{cluster}, \code{cf.em} (cell fraction computed with EM algorithm), \code{tcn.em}
(total copy number computed with EM algorithm), \code{lcn.em} (lower copy number
computed with EM algorithm), \code{ncells} (number of cells in cluster),
\code{prop.cluster} (proportion of cells per cluster), \code{gnl} (gain;neutral;loss :
1;0;-1), \code{loh} (loss of heterozygosity status), \code{state} (state of segments),
\code{cna} (whether the segment is a CNA), \code{cna_state} (state of CNA segments),
\code{prop.tot} (proportion of cells with the same state per segment),
\code{state_clonal} (state of the segment if its \code{prop.tot} is above
\code{clonal.thresh}), \code{cna_clonal} (whether the segment is a clonal CNA),
\code{cna_clonal_state} (state of clonal CNA segments).

\item \code{ncells}: Vector of number of cells per cluster.
\item \code{dipLogR.clusters}: Diploid log R ratio from the per cluster analysis.
\item \code{dipLogR.allcells}: Diploid log R ratio from the all cells analysis.
\item \code{purity.clusters}: Purity from the per cluster analysis.
\item \code{purity.allcells}: Purity from the all cells analysis.
\item \code{ploidy.clusters}: Ploidy from the per cluster analysis.
\item \code{ploidy.allcells}: Ploidy from the all cells analysis.
}
}
\description{
Performs copy number alteration (CNA) analysis on a \code{\link{muscadet}}
object by processing allelic and coverage counts across clusters and
evaluating cell fractions and copy numbers.
}
\examples{
library("facets")

# Load example muscadet object
data(muscadet_obj)

muscadet_obj <- cnaCalling(muscadet_obj,
                           omics.coverage = "ATAC",
                           depthmin.a.clusters = 3, # set low thresholds for example data
                           depthmin.c.clusters = 5,
                           depthmin.a.allcells = 3,
                           depthmin.c.allcells = 5,
                           depthmin.nor = 0)

}
\references{
Shen R, Seshan VE. FACETS: allele-specific copy number and clonal
heterogeneity analysis tool for high-throughput DNA sequencing. Nucleic
Acids Res. 2016 Sep 19;44(16):e131. doi:
\href{http://doi.org/10.1093/nar/gkw520}{10.1093/nar/gkw520}.
}
\seealso{
\code{\link[=assignClusters]{assignClusters()}}, \code{\link[=mergeCounts]{mergeCounts()}},
\code{\link[=preProcSample2]{preProcSample2()}}
}
