% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLogRatio.R
\name{computeLogRatioRNA}
\alias{computeLogRatioRNA}
\title{Compute log R ratios for scRNA-seq data}
\usage{
computeLogRatioRNA(
  matTumor,
  matRef,
  genesCoord,
  genome = "hg38",
  genesPerWindow = 101,
  refReads = 100,
  refMeanReads = 0.01,
  thresh_capping = 3,
  all_steps = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{matTumor}{Raw count matrix \emph{features x cells} for tumor/sample cells
(\code{matrix} or \code{dgCMatrix}).}

\item{matRef}{Raw count matrix \emph{features x cells} for reference cells
(\code{matrix} or \code{dgCMatrix}).}

\item{genesCoord}{Data frame of gene coordinates with columns \code{CHROM},
\code{start}, \code{end}, \code{id} (\code{data.frame}).}

\item{genome}{Reference genome name among: "hg38", "hg19" and "mm10"
(\code{character}). By default: "hg38".}

\item{genesPerWindow}{Number of genes per moving window (\code{integer} value). By
default: \code{101}.}

\item{refReads}{Minimum of reads in reference cells (\code{integer} value). By
default: \code{100}.}

\item{refMeanReads}{Minimum of average reads per reference cell (\code{integer}
value). By default: \code{0.01}.}

\item{thresh_capping}{Threshold to cap the range of log R ratio values
(\code{numeric} value). By default: \code{3}.}

\item{all_steps}{\code{TRUE} or \code{FALSE} (\code{logical}). Whether to keep intermediate
result from every step in the final object. By default: \code{FALSE}.}

\item{quiet}{Logical. If \code{TRUE}, suppresses informative messages during
execution. Default is \code{FALSE}.}
}
\value{
If \code{all_steps} is set to \code{FALSE}, a list containing:
\describe{
\item{\code{matTumor}}{Matrix of log R ratio values \emph{features x cells} for the
tumor/sample cells (\code{matrix}).}
\item{\code{matRef}}{Matrix of log R ratio values \emph{features x cells} for the
reference cells (\code{matrix}).}
\item{\code{params}}{List of parameters set for the \code{genesPerWindow}, \code{refMeans},
\code{refMeanReads} and \code{thresh_capping} arguments (\code{list}).}
\item{\code{coord}}{Data frame of coordinates for windows of peaks and associated
data along the different steps (\code{data.frame}).
Columns :
\itemize{
\item \code{CHROM}, \code{start}, \code{end}, \code{id}: coordinates and name of genes.
\item \code{sumReads.tum/ref}: sum of read counts for all cells in tumor or reference
cells.
\item \code{meanReads.tum/ref}: mean of read counts per cells for tumor or reference
cells.
\item \code{sdReads.tum/ref}: standard deviation of read counts per cells for tumor
or reference cells.
\item \code{keep}: logical, \code{TRUE} for genes to keep after filtering based on
reference coverage (depends on \code{refReads} and \code{refMeanReads} arguments).
\item \code{meanReads/sdReads.norm.tum/ref}: mean/sd of normalized counts per million
for tumor/reference cells.
\item \code{meanLRR/sdReads.raw.tum/ref}: mean/sd of raw log R ratio (LRR) for
tumor/reference cells.
\item \code{meanLRR/sdLRR.cap.tum/ref}: mean/sd of capped log R ratio (LRR) for
tumor/reference cells (depends on \code{thresh_capping} argument).
\item \code{meanLRR/sdLRR.smoo.tum/ref}: mean/sd of smoothed log R ratio (LRR) for
tumor/reference cells (means of moving windows defined by the
\code{genesPerWindow} argument).
\item \code{meanLRR/sdLRR.cent.tum/ref}: mean/sd of centered log R ratio (LRR) for
tumor/reference cells.
\item \code{meanLRR/sdLRR.corr.tum/ref}: mean/sd of final log R ratio (LRR) corrected
by reference variability for tumor/reference cells.}
}
If \code{all_steps} is set to \code{TRUE}, the previously described list can be found
for each step in a list element.
}
}
\description{
Compute log R ratios from raw count matrices with method specifically adapted
for scRNA-seq data.
}
\details{
The raw count matrix is transformed into log R ratios through the following
steps:
\itemize{
\item Match genes in count matrix with coordinates
\item Filtering on coverage (\code{refReads} and \code{refMeanReads} arguments)
\item Normalization for sequencing depth
\item Log transformation and normalization by reference data: log R ratio
\item Capping the range of values (\code{thresh_capping} argument)
\item Smoothing on genes windows
\item Centering of cells
\item Correcting by reference variability
}
}
\examples{
# Create muscomic objects
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_tumor,
  allele_counts = allele_counts_atac_tumor,
  features = peaks
)
rna <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_tumor,
  allele_counts = allele_counts_rna_tumor,
  features = genes
)
atac_ref <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_ref,
  allele_counts = allele_counts_atac_ref,
  features = peaks
)
rna_ref <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_ref,
  allele_counts = allele_counts_rna_ref,
  features = genes
)

# Create muscadet objects
muscadet <- CreateMuscadetObject(
  omics = list(atac, rna),
  bulk.lrr = bulk_lrr,
  bulk.label = "WGS",
  genome = "hg38"
)
muscadet_ref <- CreateMuscadetObject(
  omics = list(atac_ref, rna_ref),
  genome = "hg38"
)

# Compute log R ratio for RNA
obj_rna <- computeLogRatioRNA(
  matTumor = matCounts(muscadet)$RNA,
  matRef = matCounts(muscadet_ref)$RNA,
  genesCoord = coordFeatures(muscadet)$RNA,
  genome = slot(muscadet, "genome"),
  refReads = 2 # low value for example subsampled datasets
)
table(obj_rna$coord$keep)

# With results form every step when `all_steps = TRUE`
obj_rna_all <- computeLogRatioRNA(
  matTumor = matCounts(muscadet)$RNA,
  matRef = matCounts(muscadet_ref)$RNA,
  genesCoord = coordFeatures(muscadet)$RNA,
  genome = slot(muscadet, "genome"),
  refReads = 2, # low value for example subsampled datasets
  all_steps = TRUE
)
names(obj_rna_all)
table(obj_rna_all$coord$keep)
nrow(obj_rna_all$step08$matTumor)

}
\seealso{
Other computeLogRatio: 
\code{\link{computeLogRatio}()},
\code{\link{computeLogRatioATAC}()}
}
\concept{computeLogRatio}
