% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet_objects.R
\name{CreateMuscadetObject}
\alias{CreateMuscadetObject}
\title{Create a muscadet object}
\usage{
CreateMuscadetObject(
  omics,
  bulk.lrr = NULL,
  bulk.label = NULL,
  genome = "hg38"
)
}
\arguments{
\item{omics}{A list of \code{\link{muscomic}} objects (\code{list}). The names of
the list will set the names of omics in the final object, if the list is
unamed, the type is taken instead.}

\item{bulk.lrr}{A data frame containing log R ratio per genomic segments from
bulk sequencing data (\code{data.frame}). One row per segment and 4 columns
ordered as followed: chromosome (\code{integer}), start position (\code{integer}),
end position (\code{integer}), and Log R ratio value (\code{numeric}).}

\item{bulk.label}{Label for bulk data (\code{character} string).}

\item{genome}{Reference genome name among: "hg38", "hg19" and "mm10"
(\code{character} string). "hg38" by default.}
}
\value{
A \code{\link{muscadet}} object.
}
\description{
Create a \code{\link{muscadet}} object.
}
\note{
A \code{\link{muscadet}} object can contain several
\code{\link{muscomic}} objects of the same type (\code{ATAC} or \code{RNA} for slot
\code{type}) but they can't have identical labels (slot \code{label.omic}).
}
\examples{
# Create muscomic objects
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_tumor,
  allele_counts = allele_counts_atac_tumor,
  features = peaks
)
rna <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_tumor,
  allele_counts = allele_counts_rna_tumor,
  features = genes
)
atac_ref <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_ref,
  allele_counts = allele_counts_atac_ref,
  features = peaks
)
rna_ref <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_ref,
  allele_counts = allele_counts_rna_ref,
  features = genes
)

# Create muscadet objects
muscadet <- CreateMuscadetObject(
  omics = list(atac, rna),
  bulk.lrr = bulk_lrr,
  bulk.label = "WGS",
  genome = "hg38"
)
muscadet_ref <- CreateMuscadetObject(
  omics = list(atac_ref, rna_ref),
  genome = "hg38"
)

}
\seealso{
\itemize{
\item \code{\link[=CreateMuscomicObject]{CreateMuscomicObject()}}
\item \code{\link{muscomic-class}}, \code{\link{muscadet-class}}
}
}
