% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet_objects.R
\name{CreateMuscomicObject}
\alias{CreateMuscomicObject}
\title{Create a muscomic object}
\usage{
CreateMuscomicObject(
  type.omic = c("ATAC", "RNA"),
  mat_counts,
  allele_counts = NULL,
  features,
  label.omic = NULL,
  label.features = NULL
)
}
\arguments{
\item{type.omic}{Type of single cell omic, either "ATAC" or "RNA"
(\code{character} string).}

\item{mat_counts}{Matrix of raw counts \emph{features x cells} (\code{matrix} or
\code{dgCMatrix}). Rows are features, and columns are cells.}

\item{allele_counts}{Data frame of allele counts at single nucleotide
polymorphisms (SNPs) positions per cell (\code{data.frame}).}

\item{features}{Data frame of features (peaks, genes...) coordinates on
genome (\code{data.frame}).}

\item{label.omic}{Label for the single cell omic (\code{character} string). By
default "scATAC-seq" is used for "ATAC" type and "scRNA-seq" for "RNA"
type.}

\item{label.features}{Label for features (\code{character} string). By default
"peaks" is used for "ATAC" type and "genes" for "RNA" type.}
}
\value{
A \code{\link{muscomic}} object.
}
\description{
Create a \code{\link{muscomic}} object.
}
\examples{
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_tumor,
  allele_counts = allele_counts_atac_tumor,
  features = peaks
)

rna <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_tumor,
  allele_counts = allele_counts_rna_tumor,
  features = genes
)

atac_ref <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_ref,
  allele_counts = allele_counts_atac_ref,
  features = peaks
)

rna_ref <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_ref,
  allele_counts = allele_counts_rna_ref,
  features = genes
)
}
\seealso{
\itemize{
\item \code{\link[=CreateMuscadetObject]{CreateMuscadetObject()}}
\item \code{\link{muscomic-class}}, \code{\link{muscadet-class}}
}
}
