% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateMuscomicObject}
\alias{CreateMuscomicObject}
\title{Create a muscomic object}
\usage{
CreateMuscomicObject(
  type = c("ATAC", "RNA"),
  mat_counts,
  features,
  allele_counts = NULL,
  label.omic = NULL,
  label.features = NULL
)
}
\arguments{
\item{type}{Type of single cell omic, either "ATAC" or "RNA"
(\code{character} string).}

\item{mat_counts}{Matrix of raw counts \emph{features x cells} (\code{matrix} or
\code{dgCMatrix}). Rows are features (they must correspond to the id column of
\code{features}), and columns are cells.}

\item{features}{Data frame of features (peaks, genes...) coordinates on
genome (\code{data.frame}). It should contain 4 columns:
\describe{
\item{\code{CHROM}}{Chromosome names in character format, e.g. "15", "X" (\code{character}).}
\item{\code{start}}{Start positions (\code{integer}).}
\item{\code{end}}{End positions (\code{integer}).}
\item{\code{id}}{Unique identifiers, e.g. gene name "CDH1" or peak identifier
CHROM_start_end "1_1600338_1600838" (\code{character}). It should match the
feature identifiers as row names of \code{mat_counts}.}
}}

\item{allele_counts}{Data frame of allele counts at variant positions per
cell (\code{data.frame}). Variant positions can be either common single
nucleotide polymorphisms (SNPs) positions or individual-specific
heterozygous positions retrieved by bulk sequencing. The data frame format
is based on the Variant Calling Format (VCF), thereby it must contain the
following columns : \code{cell}, \code{id}, \code{CHROM}, \code{POS}, \code{REF}, \code{ALT}, \code{RD}, \code{AD},
\code{DP}, \code{GT}. See \link{allele_counts} for details.}

\item{label.omic}{Label for the single cell omic (\code{character} string). By
default "scATAC-seq" is used for "ATAC" type and "scRNA-seq" for "RNA"
type.}

\item{label.features}{Label for features (\code{character} string). By default
"peaks" is used for "ATAC" type and "genes" for "RNA" type.}
}
\value{
A \code{\link{muscomic}} object.
}
\description{
Create a \code{\link{muscomic}} object.
}
\examples{
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_tumor,
  allele_counts = allele_counts_atac_tumor,
  features = peaks
)
atac

rna <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_tumor,
  allele_counts = allele_counts_rna_tumor,
  features = genes
)
rna

atac_ref <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_ref,
  allele_counts = allele_counts_atac_ref,
  features = peaks
)

rna_ref <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_ref,
  allele_counts = allele_counts_rna_ref,
  features = genes
)
rna_ref

# without allele counts data (not required for clustering step)
atac2 <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_tumor,
  features = peaks
)
atac2

}
\seealso{
\itemize{
\item \code{\link[=CreateMuscadetObject]{CreateMuscadetObject()}}
\item \code{\link{muscomic-class}}, \code{\link{muscadet-class}}
}
}
