% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstimateGauss}
\alias{EstimateGauss}
\title{Stationary Bivariate Gaussian DPP Estimator}
\usage{
EstimateGauss(
  X,
  labels,
  lb,
  ub,
  rho1 = NA_real_,
  rho2 = NA_real_,
  alpha1 = NA_real_,
  alpha2 = NA_real_,
  estimate_alpha = TRUE
)
}
\arguments{
\item{X}{A matrix of size n x (d+1) storing the points in R^d and their label in last column.}
}
\value{
A vector with the estimated model parameters.
}
\description{
This function estimates the parameters of a stationary bivariate Gaussian
DPP from a set of observed points and labels.
}
\examples{
dpp <- sim_gauss5[[1]]
X <- cbind(dpp$x, dpp$y)
labels <- dpp$marks
rho1 <- rho2 <- 100
rho12 <- sqrt(0.5 * sqrt(rho1 * rho2)
alpha1 <- alpha2 <- 0.03
alpha12 <- 0.03
d <- 2
EstimateGauss(
  X = X,
  labels = labels,
  lb = rep(-0.5, ncol(X)),
  ub = rep( 0.5, ncol(X)),
  rho1 = rho1,
  rho2 = rho2,
  alpha1 = alpha1,
  alpha2 = alpha2,
  estimate_alpha = FALSE
)
}
