% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{genome_chrom}
\alias{genome_chrom}
\alias{hg38_chrom}
\alias{hg19_chrom}
\alias{mm10_chrom}
\title{Genome chromosome sizes (internal data)}
\format{
A \code{\link{GRanges}} object containing:
\describe{
\item{\code{seqnames}}{Chromosome name: 1 to 22, X and Y for human ; and 1 to
19, X and Y for mouse (\code{Rle}).}
\item{\code{ranges}}{Ranges of chromosomes (\code{IRanges}).}
\item{\code{strand}}{Strand information * (\code{Rle}).}
}
}
\description{
\code{\link{GRanges}} objects containing chromosomes sizes for
hg38, hg19 and mm10 genome assemblies.
}
\note{
Data obtained from assemblies provided by the \link{BSgenome} package.
\itemize{
\item \code{BSgenome.Hsapiens.UCSC.hg38} version 1.4.5 - \code{GRCh38.p14}
\item \code{BSgenome.Hsapiens.UCSC.hg19} version 1.4.3 - \code{GRCh37.p13}
\item \code{BSgenome.Mmusculus.UCSC.mm10} version 1.4.3 - \code{GRCm38.p6}
}
}
\examples{
muscadet:::hg38_chrom
muscadet:::hg19_chrom
muscadet:::mm10_chrom
}
\keyword{internal}
