% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLogRatio.R
\name{getLogRatioBulk}
\alias{getLogRatioBulk}
\title{Retrieve log R ratio from Bulk data on single-cell features (internal)}
\usage{
getLogRatioBulk(x, bulk.lrr)
}
\arguments{
\item{x}{A \code{muscomic} object containing the single-cell omic data, which
includes the feature coordinates (\code{muscomic}).}

\item{bulk.lrr}{A data frame containing log R ratio per genomic segments from
bulk sequencing data (\code{data.frame}). One row per segment and 4 columns
ordered as followed: chromosome (\code{character}), start position (\code{integer}),
end position (\code{integer}), and Log R ratio value (\code{numeric}).}
}
\value{
A data frame of single-cell omic features, with columns corresponding
to \code{CHROM}, \code{start}, \code{end}, and \code{bulk.lrr}, where \code{bulk.lrr} corresponds to
the log R ratio values retrieved from bulk data matching the feature
coordinates.
}
\description{
This internal function assigns the log R ratio (LRR) values from bulk data
segments to single-cell omic features. It matches the bulk data segments to
corresponding genomic features from the single-cell omic data and returns a
table of single-cell features with the corresponding bulk LRR values.
}
\examples{
# muscomic object inside a muscadet object
muscadet_obj$ATAC

# Bulk Log R ratio data
head(muscadet_obj$bulk.data$logratio)

features_bulk_lrr <- getLogRatioBulk(
  x = muscadet_obj$ATAC,
  bulk.lrr = muscadet_obj$bulk.data$logratio
)
head(features_bulk_lrr)


}
\seealso{
\itemize{
\item example data: \code{\link{bulk_lrr}}
\item functions: \code{\link{muscomic}}, \code{\link{muscadet}}
}
}
