% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmapMuscadet}
\alias{heatmapMuscadet}
\title{Heatmap Plot Function for muscadet Object}
\usage{
heatmapMuscadet(
  x,
  filename,
  k = NULL,
  clusters = NULL,
  title = "",
  add_bulk_lrr = TRUE,
  show_missing = TRUE,
  white_scale = c(0.3, 0.7),
  colors = NULL,
  quiet = FALSE
)
}
\arguments{
\item{x}{A muscadet object containing LRR data for all omics (with
\code{\link[=ComputeLogRatio]{ComputeLogRatio()}}) and clustering data (with
\code{\link[=clusterMuscadet]{clusterMuscadet()}}) (\code{muscadet}).}

\item{filename}{Character string specifying the file path to save the heatmap
image in the PNG format (\code{character} string).}

\item{k}{(Optional) Integer specifying the cluster number to plot
(\code{integer}). It should be within the range of k used for clustering (with
\code{\link[=clusterMuscadet]{clusterMuscadet()}}). Should be provided if \code{clusters} is \code{NULL}.}

\item{clusters}{(Optional) A custom named vector of cluster assignments
(\code{integer} named vector). Names must corresponds the cell names of the
\code{muscadet} object: to the totality of cells if \code{show_missing = TRUE}, or to
only common cells (cells with data in all omics) if \code{show_missing = FALSE}.
Should be provided if \code{k} is \code{NULL}.}

\item{title}{Character string for the title of the heatmap (\code{character} string).}

\item{add_bulk_lrr}{Logical value indicating whether to add bulk LRR
data as annotations if present in the muscadet object (\code{logical}). Default is \code{TRUE}.}

\item{show_missing}{Logical value indicating whether to show missing cells
(cells with missing data in at least one omic) in the heatmaps (\code{logical}). Default is \code{TRUE}.}

\item{white_scale}{Numeric vector of length 2 with values between 0 and 1,
defining the white color boundaries (\code{numeric} vector). This parameter
specifies the quantiles of the LRR reference data that will define
the boundaries for the white color in the heatmap. LRR values falling
within this range are considered close enough to the majority of the LRR
reference data, indicating no significant gain or loss of coverage, thereby
are represented as white on the heatmap.}

\item{colors}{Vector of colors for the cluster annotation. Default is \code{NULL}, which uses predefined colors.}

\item{quiet}{Logical value indicating whether to suppress messages. Default is \code{FALSE}.}
}
\value{
A heatmap plot object saved as a png image using the provided file name.
}
\description{
This function generates a heatmap to visualize log R ratio (LRR) data
contained in muscadet objects. One heatmap is generated per omic, rows are
cells and columns are chromosomes, for muscadet object containing multiple
omics, the heatmaps are plotted horizontally aligned. The cells can be
clustered for a specific k number of cluster following the clustering step of
muscadet object, or custom cluster assignments can be used. Additionally, LRR
values from bulk sequencing data can be plotted as an annotation under the
heatmaps.
}
\examples{
# Load example muscadet object
data(muscadet_obj)

# Perform clustering (if not already done)
muscadet_obj <- clusterMuscadet(muscadet_obj, k_range = 2:5)

# Custom title with clustering parameters
title <- paste("Example sample |",
     slot(muscadet_obj, "clustering")[["params"]][["dist_method"]],
     slot(muscadet_obj, "clustering")[["params"]][["hclust_method"]], "|",
     "k=3", "|",
     "Weights of omics:",
     paste(slot(muscadet_obj, "clustering")[["params"]][["weights"]], collapse = ", "))

# Generate heatmap
heatmapMuscadet(
    muscadet_obj,
    k = 3,
    filename = file.path(getwd(), "heatmap_muscadet_k3.png"),
    title = title
)

heatmapMuscadet(
    muscadet_obj,
    k = 3,
    filename = file.path(getwd(), "heatmap_muscadet_k3_commoncells.png"),
    title = title,
    show_missing = FALSE
)

}
