% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmapMuscadet}
\alias{heatmapMuscadet}
\title{Heatmap plot for \code{muscadet} object}
\usage{
heatmapMuscadet(
  x,
  filename = NULL,
  k = NULL,
  clusters = NULL,
  title = "",
  add_bulk_lrr = NULL,
  show_missing = TRUE,
  row_annots = NULL,
  white_scale = c(0.3, 0.7),
  colors = NULL,
  png_res = 300,
  quiet = FALSE
)
}
\arguments{
\item{x}{A \code{\link{muscadet}} object containing LRR data for all omics
(with \code{\link[=computeLogRatio]{computeLogRatio()}}) and clustering data (with
\code{\link[=clusterMuscadet]{clusterMuscadet()}}) (\code{muscadet}).}

\item{filename}{(Optional) Character string specifying the file path to save
the heatmap image in the PNG (if it ends by .png) or PDF (if it ends by
.pdf) format (\code{character} string).}

\item{k}{(Optional) Integer specifying the cluster number to plot
(\code{integer}). It should be within the range of k used for clustering (with
\code{\link[=clusterMuscadet]{clusterMuscadet()}}). Should be provided if \code{clusters} is \code{NULL}.}

\item{clusters}{(Optional) A custom named vector of cluster assignments
(\code{integer} named vector). Names must corresponds to cell names within the
muscadet object \code{x}. If it contains less cells than the muscadet object
\code{x}, the missing cells are filtered out and not displayed in the heatmap.
If \code{show_missing = FALSE} only the provided cells with data in all omics
will be displayed. Should be provided if \code{k} is \code{NULL}.}

\item{title}{Character string for the title of the heatmap (\code{character}
string). Default is an empty character string.}

\item{add_bulk_lrr}{Logical value indicating whether to add bulk LRR data as
annotations if present in the muscadet object (\code{logical}). By default, it
is set to \code{NULL}, which will automatically include bulk LRR data if it is
available in the muscadet object.}

\item{show_missing}{Logical value indicating whether to show missing cells
(cells with missing data in at least one omic) in the heatmaps (\code{logical}).
Default is \code{TRUE}.}

\item{row_annots}{Optional. A list of \linkS4class{HeatmapAnnotation} objects from
the \link{ComplexHeatmap} package, specifying row annotations to add on left
part of the heatmap. Each element in the list must be of class
(HeatmapAnnotation)\linkS4class{HeatmapAnnotation}, must be a row annotation
(using \verb{[rowAnnotation()]} or \verb{[HeatmapAnnotation()]} with
\code{which = 'row'}), and must have a unique name (\code{name} argument in
\verb{[rowAnnotation()]} or \verb{[HeatmapAnnotation()]}). By default is \code{NULL}, no
row annotations is added.}

\item{white_scale}{Numeric vector of length 2 or a list of numeric vectors
(\code{numeric} vector or \code{list}).
\itemize{
\item If a numeric vector of length 2, the same white color boundaries are
applied to all omics in the muscadet object. E.g. \code{c(0.3, 0.7)} (default).
\item If a list (named or not with omics name), it must have the same length as
the number of omics in the muscadet object, where each vector element
applies the white color boundaries for a specific omic. E.g. \code{list(c(0.3, 0.7), c(0.4, 0.6))} uses 0.3 and 0.7 quantiles of LRR ref data for the 1st
omic heatmap, and 0.4 and 0.6 quantiles for the second.

Values of the vectors must be between 0 and 1, specifying the quantiles of
the LRR reference data that define the boundaries for the white color in
each heatmap. LRR values falling within this range are considered close to
the majority of the LRR reference data, indicating no significant gain or
loss of coverage, and are represented as white on the heatmap. Default is
\code{c(0.3, 0.7)}.
}}

\item{colors}{Vector of colors for the cluster annotation (\code{character}
vector). Default is \code{NULL}, which uses predefined colors.}

\item{png_res}{Resolution in ppi for \code{\link[grDevices:png]{grDevices::png()}} if \code{filename} ends
with the .png extension (\code{numeric}). Default is \code{300}.}

\item{quiet}{Logical value indicating whether to suppress messages. Default
is \code{FALSE}.}
}
\value{
A list containing:
\itemize{
\item \code{plot}: A \code{\link{gTree}} object created with \code{\link[grid:grid.grab]{grid::grid.grab()}} (\code{\link{gTree}}).
\item \code{width}: Width of the heatmap plot in mm (\code{\link{unit}}).
\item \code{height}: Height of the heatmap plot in mm (\code{\link{unit}}).
}

If the \code{filename} argument is provided, the heatmap is directly saved as a
PNG image at the provided path.
}
\description{
This function generates a heatmap to visualize log R ratio (LRR) data
contained in \code{\link{muscadet}} objects. One heatmap is generated per
omic, rows are cells and columns are chromosomes, for \code{muscadet} object
containing multiple omics, the heatmaps are plotted horizontally aligned. The
cells can be clustered for a specific k number of cluster following the
clustering step of \code{muscadet} object, or custom cluster assignments can be
used. Additionally, LRR values from bulk sequencing data can be plotted as an
annotation under the heatmaps.
}
\examples{
# Load example muscadet object
data(muscadet_obj)

# Perform clustering (if not already done)
muscadet_obj <- clusterMuscadet(muscadet_obj, k_range = 2:5)

# Custom title with clustering parameters
title <- paste("Example sample |",
     slot(muscadet_obj, "clustering")[["params"]][["dist_method"]],
     slot(muscadet_obj, "clustering")[["params"]][["hclust_method"]], "|",
     "k=3", "|",
     "Weights of omics:",
     paste(slot(muscadet_obj, "clustering")[["params"]][["weights"]], collapse = ", "))

# Generate heatmap
heatmapMuscadet(
    muscadet_obj,
    k = 3,
    filename = file.path("heatmap_muscadet_k3.png"),
    title = title
)
heatmapMuscadet(
    muscadet_obj,
    k = 3,
    filename = file.path("heatmap_muscadet_k3_commoncells.png"),
    title = title,
    show_missing = FALSE
)

# Loop over k
for (k in names(muscadet_obj$clustering$clusters)) {
    filename <- paste0("heatmap_muscadet_k", k, ".png")
    title <- paste(
        "Example sample |",
        muscadet_obj$clustering$params[["dist_method"]],
        muscadet_obj$clustering$params[["hclust_method"]],
        "|",
        paste0("k=", k),
        "|",
        "Weights of omics:",
        paste(muscadet_obj$clustering$params[["weights"]], collapse = ", ")
    )
    heatmapMuscadet(
        muscadet_obj,
        k = as.integer(k),
        filename = filename,
        title = title
    )
}

# Add row annotation

# Define example annotation
muscadet_cells <- sort(Reduce(union, Cells(muscadet_obj)))
cells_origin <- setNames(c(
    rep("sample1", ceiling(length(muscadet_cells) / 2)),
    rep("sample2", floor(length(muscadet_cells) / 2))
    ),
    muscadet_cells
)

# Create row annotation
ha <- rowAnnotation(
    annot = anno_simple(
        cells_origin,
        col = c(
            "sample1" = "cadetblue3",
            "sample2" = "orchid3"
        )
    ),
    name = "origin",
    annotation_label = "origin",
    annotation_name_gp = gpar(fontsize = 10)
)

# Generate heatmap with supplementary row annotation
heatmapMuscadet(
    muscadet_obj,
    k = 3,
    filename = file.path("heatmap_muscadet_k3.png"),
    title = title,
    row_annots = list(ha)
)

}
