% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeCounts}
\alias{mergeCounts}
\title{Merge counts for \code{muscadet} objects}
\usage{
mergeCounts(x, reference)
}
\arguments{
\item{x}{A \code{\link{muscadet}} object containing sample data (\code{muscadet}).
This object must include clustering assignments in the
\code{cnacalling$clusters} slot.}

\item{reference}{A \code{\link{muscadet}} object containing reference data
(\code{muscadet}).}
}
\value{
A modified \code{\link{muscadet}} object corresponding to the \code{x} muscadet object,
with updated \code{cnacalling} slot containing:
\itemize{
\item \code{allelic.counts}: Processed allelic counts on SNP positions, for all omics
\item \code{coverage.counts}: Processed coverage counts merged with the reference.
\item \code{combined.counts}: Combined data for allelic and coverage counts.
}
Abbreviations:
\itemize{
\item RD = Reference allele read depth
\item AD = Alternative allele read depth
\item DP = Total read depth
\item TUM = tumor sample
\item NOR = normal reference
\item omic = omic specific (\code{omic} column)
\item all = for all omics
}
}
\description{
This function combines allelic and coverage counts from all omics per cluster
for both sample and reference. The resulting merged data is
stored in the \code{cnacalling} slot of the sample \code{muscadet} object.
}
\examples{
# Load example muscadet objects
data(muscadet_obj)
data(muscadet_obj_ref)

# Merge counts from all omics from both sample and reference
muscadet_obj <- mergeCounts(muscadet_obj, muscadet_obj_ref)

}
