% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle-dpp}
\alias{mle-dpp}
\alias{mle_dpp_gauss}
\alias{mle_dpp_bessel}
\title{Maximum Likelihood Estimator of Stationary Bivariate DPPs}
\usage{
mle_dpp_gauss(X, labels, lb = rep(-0.5, ncol(X)), ub = rep(0.5,
  ncol(X)), rho1 = NA, alpha1 = NA, rho2 = NA, alpha2 = NA,
  estimate_alpha = TRUE)

mle_dpp_bessel(X, labels, lb = rep(-0.5, ncol(X)), ub = rep(0.5,
  ncol(X)), rho1 = NA, alpha1 = NA, rho2 = NA, alpha2 = NA,
  estimate_alpha = TRUE)
}
\arguments{
\item{X}{An n x d matrix storing n observed points in R^d.}

\item{labels}{An n-dimensional integer vector storing the labels of each
observed point.}

\item{lb}{A d-dimensional numeric vector storing the lower bounds of the
spatial domain (default: \code{rep(-0.5, ncol(X))}).}

\item{ub}{A d-dimensional numeric vector storing the upper bounds of the
spatial domain (default: \code{rep( 0.5, ncol(X))}).}

\item{rho1}{Value of the first intensity. If set, it requires toset also the
\code{alpha1} parameter.}

\item{alpha1}{Value of the first alpha If set to \code{NA} (default), the
parameter is estimated.}

\item{rho2}{Value of the second intensity. If set, it requires to set also
the \code{alpha2} parameter.}

\item{alpha2}{Value of the second alpha If set to \code{NA} (default), the
parameter is estimated.}

\item{estimate_alpha}{A boolean specifying whether the marginal alpha's
should be estimated (default: \code{TRUE}).}
}
\value{
A list as output from \code{\link[stats]{optim}}.
}
\description{
Maximum Likelihood Estimator of Stationary Bivariate DPPs
}
\examples{
dpp <- bessel_tauSq0p5[[1]]
X <- cbind(dpp$x, dpp$y)
labels <- dpp$marks
rho1 <- rho2 <- 100
rho12 <- sqrt(0.5) * sqrt(rho1 * rho2)
alpha1 <- alpha2 <- 0.03
alpha12 <- 0.03
d <- 2
mle_dpp_bessel(
  X = X,
  labels = labels,
  rho1 = rho1,
  rho2 = rho2,
  alpha1 = alpha1,
  alpha2 = alpha2,
  estimate_alpha = FALSE
)
}
