% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Estimation of multi-mark DPP via Maximum Likelihood}
\usage{
mle(
  data,
  initial_guess = NULL,
  fixed_marginal_parameters = FALSE,
  model = "gauss",
  optimizer = "bobyqa",
  num_threads = 1,
  N = 512,
  verbose_level = 0
)
}
\arguments{
\item{data}{A \code{\link[spatstat]{ppp}} object storing an observed planar
determinantal point process.}

\item{initial_guess}{A vector providing an intial guess for the model
parameters that maximize the likelihood. Defaults to \code{NULL}, which
initializes at \code{0.5} all parameters after suitable transformation into
\link{0,1}. If provided, expected order is \code{c(alpha1, alpha2, alpha12, tau)}.}

\item{fixed_marginal_parameters}{Boolean specifying whether the marginal
parameters should be estimated separately using the marginal likelihood and
then fixed to further estimate the cross parameters. Default is \code{FALSE}.}

\item{model}{A DPP model. For now either gauss or bessel (default: gauss).}

\item{optimizer}{Any derivative-free optimizer in NlOpt (default: bobyqa).}

\item{num_threads}{Number of threas to run on (default: 1).}

\item{N}{Maximum truncation index for Fourier transform (default: 512).}

\item{verbose_level}{Display information during optimization. Choose \code{0}
for no information, \code{1} for global information at likelihood setup or
\code{2} for detailed  information at each function evaluation. Defaults to
\code{0}.}
}
\value{
A list with two components: \code{par} (optimized model parameters)
and \code{value} (\code{-2 logLik} at the maximum likelihood).
}
\description{
Estimation of multi-mark DPP via Maximum Likelihood
}
\examples{
opt <- mle(sim_gauss5[[1]])
}
