% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{[.muscadet}
\alias{[.muscadet}
\alias{muscadet-access}
\alias{[.muscomic}
\alias{$.muscadet}
\alias{$.muscomic}
\alias{[<-.muscadet}
\alias{[<-.muscomic}
\alias{$<-.muscadet}
\alias{$<-.muscomic}
\title{Access and assignment methods for \code{muscadet} objects}
\usage{
\method{[}{muscadet}(x, i, ...)

\method{[}{muscomic}(x, i, ...)

\method{$}{muscadet}(x, name)

\method{$}{muscomic}(x, name)

\method{[}{muscadet}(x, i) <- value

\method{[}{muscomic}(x, i) <- value

\method{$}{muscadet}(x, i) <- value

\method{$}{muscomic}(x, i) <- value
}
\arguments{
\item{x}{A \code{\link[=muscadet-class]{muscadet}} or \code{\link[=muscomic-class]{muscomic}}
object.}

\item{i}{The name of the slot (or omic).}

\item{...}{Other arguments (ignored).}

\item{name}{The name of the slot (or omic).}

\item{value}{The new value to assign.}
}
\value{
The selected slot or the omic dataset (\code{[muscadet::muscomic()]}
object) for \code{muscadet} objects. The selected slot for \code{muscomic} objects.

The updated \code{[muscadet::muscadet()]} or
\code{[muscadet::muscomic()]} object.
}
\description{
Simplified access to omic datasets and slots inside
\code{\link{muscadet}} objects.

Assign new data in a \code{[muscadet::muscadet()]} or
\code{[muscadet::muscomic()]} object. For \code{muscadet} objects, the omic datasets in
the \code{omics} slot can be directly reassigned.
}
\examples{
# Load example muscadet object
# data("exdata_muscadet")

# Access to muscadet omics or slots
exdata_muscadet["ATAC"]
exdata_muscadet["genome"]

# Load example muscadet object
# data("exdata_muscadet")

# Access to muscomic slots
exdata_muscadet["ATAC"]["label.omic"]

# Load example muscadet object
# data("exdata_muscadet")

# Access to muscadet omics or slots
exdata_muscadet$ATAC
exdata_muscadet$genome

# Load example muscadet object
# data("exdata_muscadet")

# Access to muscomic slots
exdata_muscadet$ATAC$label.omic

# Load example muscadet object
# data("exdata_muscadet")

# Assign new data in muscadet object
exdata_muscadet["genome"] <- "hg38"

# Load example muscadet object
# data("exdata_muscadet")

# Assign new data in muscomic object
exdata_muscadet["ATAC"]["label.omic"] <- "scATAC-seq"

# Load example muscadet object
# data("exdata_muscadet")

# Assign new data in muscadet object
exdata_muscadet$genome <- "hg38"

# Load example muscadet object
# data("exdata_muscadet")

# Assign new data in muscomic object
exdata_muscadet$ATAC$label.omic <- "scATAC-seq"

}
