% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{muscadet-methods}
\alias{muscadet-methods}
\alias{muscomic-methods}
\alias{coordFeatures}
\alias{matCounts}
\alias{matLogRatio}
\alias{Cells.muscomic}
\alias{Cells.muscadet}
\alias{Features.muscomic}
\alias{Features.muscadet}
\alias{coordFeatures,muscomic-method}
\alias{coordFeatures,muscadet-method}
\alias{matCounts,muscomic-method}
\alias{matCounts,muscadet-method}
\alias{matLogRatio,muscomic-method}
\alias{matLogRatio,muscadet-method}
\title{Methods for \code{[muscadet::muscomic()]} and \code{[muscadet::muscadet()]}
objects}
\usage{
coordFeatures(x)

matCounts(x)

matLogRatio(x)

\method{Cells}{muscomic}(x, ...)

\method{Cells}{muscadet}(x, ...)

\method{Features}{muscomic}(x, ...)

\method{Features}{muscadet}(x, ...)

\S4method{coordFeatures}{muscomic}(x)

\S4method{coordFeatures}{muscadet}(x)

\S4method{matCounts}{muscomic}(x)

\S4method{matCounts}{muscadet}(x)

\S4method{matLogRatio}{muscomic}(x)

\S4method{matLogRatio}{muscadet}(x)
}
\arguments{
\item{x}{A \code{[muscadet::muscomic()]} or \code{[muscadet::muscadet()]}
object.}

\item{...}{Other arguments (ignored).}
}
\value{
\code{Cells}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a vector of cell names.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a vector of cell names (one omic),
or a list of vectors of cell names, one list element per omic.
}

\code{Features}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a vector of feature names.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a vector of feature names (one omic),
or a list of vectors of feature names, one list element per omic.
}

\code{coordFeatures}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a data frame of feature
coordinates.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a data frame of feature
coordinates (one omic), or a list of data frames of feature coordinates,
with one list element per omic.
}

\code{matCounts}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}} \emph{features x cells}.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}}
\emph{features x cells} (one omic), or a list of \code{\link[Matrix:dgCMatrix-class]{dgCMatrix}},
with one list element per omic.
}

\code{matLogRatio}:
\itemize{
\item if \code{x} is a \code{\link[=muscomic]{muscomic()}} object: a \code{matrix} \emph{features x cells}.
\item if \code{x} is a \code{\link[=muscadet]{muscadet()}} object: a \code{matrix}
\emph{features x cells} (one omic), or a list of \code{matrix},
with one list element per omic.
}
}
\description{
Methods to facilitate access to data within the \code{[muscadet::muscomic()]}
and \code{[muscadet::muscadet()]} objects.
\itemize{
\item \code{Cells()}: Get cell identifiers (addition of methods for \code{muscomic} and
\code{muscadet} to \code{\link[SeuratObject:Cells]{SeuratObject::Cells()}}).
\item \code{Features()}: Get feature identifiers (addition of methods for \code{muscomic}
and \code{muscadet} to \code{\link[SeuratObject:Cells]{SeuratObject::Features()}}).
\item \code{coordFeatures()}: Get coordinates of features data frames.
\item \code{matCounts()}: Get raw count matrices.
\item \code{matLogRatio()}: Get log R ratio matrices.
}
}
\examples{
library("SeuratObject")

# Load example muscadet object
# data("muscadet_obj")

Cells(muscadet_obj) # list of 2 cell names vectors for the 2 omics
Cells(muscadet_obj)$ATAC # cell names vector from the omic ATAC
Cells(muscadet_obj$ATAC) # cell names vector from the ATAC muscomic object

}
\seealso{
\code{[muscadet::muscomic-class()]}
\code{[muscadet::muscadet-class()]}
}
