% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet_objects.R
\name{muscadet-methods}
\alias{muscadet-methods}
\alias{muscomic-methods}
\alias{Cells}
\alias{Features}
\alias{coordFeatures}
\alias{matCounts}
\alias{matLogRatio}
\alias{Cells,muscomic-method}
\alias{Cells,muscadet-method}
\alias{Features,muscomic-method}
\alias{Features,muscadet-method}
\alias{coordFeatures,muscomic-method}
\alias{coordFeatures,muscadet-method}
\alias{matCounts,muscomic-method}
\alias{matCounts,muscadet-method}
\alias{matLogRatio,muscomic-method}
\alias{matLogRatio,muscadet-method}
\title{\code{\link{muscomic}} and \code{\link{muscadet}} methods}
\usage{
Cells(x, ...)

Features(x, ...)

coordFeatures(x, ...)

matCounts(x, ...)

matLogRatio(x, ...)

\S4method{Cells}{muscomic}(x)

\S4method{Cells}{muscadet}(x)

\S4method{Features}{muscomic}(x)

\S4method{Features}{muscadet}(x)

\S4method{coordFeatures}{muscomic}(x)

\S4method{coordFeatures}{muscadet}(x)

\S4method{matCounts}{muscomic}(x)

\S4method{matCounts}{muscadet}(x)

\S4method{matLogRatio}{muscomic}(x)

\S4method{matLogRatio}{muscadet}(x)
}
\arguments{
\item{x}{A \code{\link{muscomic}} or \code{\link{muscadet}} object.}
}
\value{
\code{Cells}:
\itemize{
\item \code{muscomic}: A vector of cell names.
\item \code{muscadet}: A list of vectors of cell names, one list element per omic.
}

\code{Features}:
\itemize{
\item \code{muscomic}: A vector of feature names.
\item \code{muscadet}: A list of vectors of feature names, one list element per omic.
}

\code{coordFeatures}:
\itemize{
\item \code{muscomic}: A data frame of feature coordinates.
\item \code{muscadet}: A list of data frames of feature coordinates, one list element per omic.
}

\code{matCounts}:
\itemize{
\item \code{muscomic}: A \code{\link{dgCMatrix}} \emph{features x cells}.
\item \code{muscadet}: A list of \code{\link{dgCMatrix}} \emph{features x cells}, one list element per omic.
}

\code{matLogRatio}:
\itemize{
\item \code{muscomic}: A \code{\link{matrix}} \emph{features x cells}.
\item \code{muscadet}: A list of \code{\link{matrix}} \emph{features x cells}, one list element per omic.
}
}
\description{
Methods for \code{\link{muscomic}} and \code{\link{muscadet}} objects, to
facilitate access to slots within the objects.
}
\seealso{
\code{\link{muscomic-class}} \code{\link{muscadet-class}}
}
