% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet-package.R
\docType{package}
\name{muscadet-package}
\alias{muscadet-package}
\title{muscadet: Multiomics Single-cell Copy number Alterations Detection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Identification of copy number alterations (CNAs) in cancer cells from single-cell multiomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ICAGEN/muscadet}
  \item Report bugs at \url{https://github.com/ICAGEN/muscadet/issues}
}

}
\author{
\strong{Maintainer}: Marie Denoulet \email{marie.denoulet@inserm.fr} (\href{https://orcid.org/0000-0002-9668-0439}{ORCID})

Authors:
\itemize{
  \item Céline Vallot \email{celine.vallot@curie.fr} (\href{https://orcid.org/0000-0003-1601-2359}{ORCID})
  \item Eric Letouzé \email{eric.letouze@inserm.fr} (\href{https://orcid.org/0000-0002-6369-2839}{ORCID})
}

}
\keyword{internal}
