% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{rbidpp}
\alias{rbidpp}
\title{Random generation of point patterns from a cross-type DPP}
\usage{
rbidpp(
  n = 1L,
  rho1 = 100,
  rho2 = 100,
  alpha1 = 0.03,
  alpha2 = 0.03,
  alpha12 = 0.035,
  tau = 0.5,
  L = 1,
  d = 2L,
  model = c("Gauss", "Bessel"),
  max_trunc = 1000L
)
}
\arguments{
\item{n}{An integer value specifying the sample size. Defaults to \code{1L}.}

\item{rho1}{A numeric value specifying the intensity of the 1st marginal DPP.
Defaults to \code{100L}.}

\item{rho2}{A numeric value specifying the intensity of the 2nd marginal DPP.
Defaults to \code{100L}.}

\item{alpha1}{A numeric value specifying the repulsion rate of the 1st DPP.
Defaults to \code{0.03}.}

\item{alpha2}{A numeric value specifying the repulsion rate of the 2nd DPP.
Defaults to \code{0.03}.}

\item{alpha12}{A numeric value specifying the repulsion rate between marks.
Defaults to \code{0.05}.}

\item{tau}{A numeric value specifying the correlation between marks. Defaults
to \code{0.2}.}

\item{L}{An integer value specifying the window size of the point pattern.
Defaults to \code{1L}.}

\item{d}{An integer value specifying the dimension of the DPP. Defaults to
\code{2L}.}

\item{model}{A string specifying the model to be used. Choices are \code{"Gauss"}
or \code{"Bessel"}. Defaults to \code{"Gauss"}.}

\item{max_trunc}{An integer value specifying the maximum number of basis
functions in the Fourier decomposition. Defaults to \code{1000L}.}
}
\value{
An object of class \link[spatstat.geom:ppp]{spatstat.geom::ppp} storing the simulated point
pattern.
}
\description{
Simulates one or more planar point patterns following a given multi-mark DPP.
}
\keyword{internal}
