% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{show,muscadet-method}
\alias{show,muscadet-method}
\title{muscadet object overview}
\usage{
\S4method{show}{muscadet}(object)
}
\arguments{
\item{object}{A \code{[muscadet::muscadet()]} object.}
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly returns
\code{NULL}
}
\description{
Overview of a \code{[muscadet::muscadet()]} object.
}
\examples{
# Load example muscadet object
# data("muscadet_obj")

# Overview of the muscadet object
show(muscadet_obj)

# Overview of the muscomic objects within
show(slot(muscadet_obj, "omics"))

# Overview of the first muscomic objects within
show(slot(muscadet_obj, "omics")[[1]])

}
\seealso{
\code{[muscadet::muscadet-class()]}, \code{\link[=CreateMuscadetObject]{CreateMuscadetObject()}}
}
\keyword{internal}
