% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscadet_objects.R
\name{show,muscadet-method}
\alias{show,muscadet-method}
\title{muscadet object overview}
\usage{
\S4method{show}{muscadet}(object)
}
\arguments{
\item{object}{A \code{\link{muscadet}} object.}
}
\value{
Prints summary to \code{\link[base]{stdout}} and invisibly returns
\code{NULL}
}
\description{
Overview of a \code{\link{muscadet}} object.
}
\examples{
atac <- CreateMuscomicObject(
  type = "ATAC",
  mat_counts = mat_counts_atac_tumor,
  allele_counts = allele_counts_atac_tumor,
  features = peaks
)
rna <- CreateMuscomicObject(
  type = "RNA",
  mat_counts = mat_counts_rna_tumor,
  allele_counts = allele_counts_rna_tumor,
  features = genes
)

muscadet <- CreateMuscadetObject(
  omics = list(atac, rna),
  bulk_lrr = bulk_lrr,
  bulk.label = "WGS",
  genome = "hg38"
)
}
\seealso{
\code{\link{muscadet-class}}, \code{\link[=CreateMuscadetObject]{CreateMuscadetObject()}}
}
\keyword{internal}
